% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{build_ards_df}
\alias{build_ards_df}
\title{Build dataframe with ARDS by imaging procedures}
\usage{
build_ards_df(trial_data, pid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}
}
\value{
dataframe, which contains the ards data
}
\description{
Build dataframe with ARDS by imaging procedures
}
