% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{build_hap_pcs_score_df_without_proms}
\alias{build_hap_pcs_score_df_without_proms}
\title{Create data frame with Post-COVID variables without PROMs in Post-COVID-Score}
\usage{
build_hap_pcs_score_df_without_proms(trial_data, pid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in data}
}
\value{
A dataframe with Post-COVID variables without PROMs in Post-COVID-Score
}
\description{
returns a dataframe, which contains the Post-COVID variables without PROMs in Post-COVID-Score
}
