% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{build_suep_long_symptom_df}
\alias{build_suep_long_symptom_df}
\title{Create long symptom data frame}
\usage{
build_suep_long_symptom_df(trial_data, pid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in data}
}
\value{
A dataframe with one row per symptom per patient
}
\description{
returns a dataframe, which contains the 'long_symptom_data'
}
