% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{build_sy_parents_children_merged_df}
\alias{build_sy_parents_children_merged_df}
\title{Build symptom subform data frame (merged with parent and cild forms)}
\usage{
build_sy_parents_children_merged_df(trial_data, pid)
}
\arguments{
\item{trial_data}{A SecuTrial data object}

\item{pid}{column name of patient ID in data}
}
\description{
Build symptom subform data frame (merged with parent and cild forms)
}
