% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{build_sy_subform_merged_df}
\alias{build_sy_subform_merged_df}
\title{Build symptom subform data frame}
\usage{
build_sy_subform_merged_df(symptom_parentformtablenames, trial_data, pid)
}
\arguments{
\item{symptom_parentformtablenames}{A vector containing parent table names for symptoms}

\item{trial_data}{A SecuTrial data object}

\item{pid}{column name of patient ID in data}
}
\description{
Build symptom subform data frame
}
