% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{build_who_scale_hap}
\alias{build_who_scale_hap}
\title{Calculate and categorize WHO-Scale}
\usage{
build_who_scale_hap(trial_data, pid, table_names)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}

\item{table_names}{A vector containing table names of data export}
}
\description{
adds the following columns to eosfci
ecu_who_scale.factor, ecu_who_scale
}
\details{
adds the following columns to osfci
ecu_who_scale_max.factor, ecu_who_scale_max
}
