% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{build_who_scale_suep_df}
\alias{build_who_scale_suep_df}
\title{Calculate and categorize WHO-Scale}
\usage{
build_who_scale_suep_df(trial_data, pid, docid, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}

\item{docid}{column name of document ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\value{
dataframe, which contains the who-scale per visit
}
\description{
Calculate and categorize WHO-Scale
}
