% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_6ils_total}
\alias{calculate_6ils_total}
\title{Calculate Six Item Loneliness Scale (6 ILS) sum score}
\usage{
calculate_6ils_total(
  six_ils1,
  six_ils2,
  six_ils3,
  six_ils4,
  six_ils5,
  six_ils6
)
}
\arguments{
\item{six_ils1}{vector for item "I miss having a really close friend."}

\item{six_ils2}{vector for item "There are plenty of people I can lean on when I have problems ."}

\item{six_ils3}{vector for item "There are many people I can trust completely."}

\item{six_ils4}{vector for item "I miss the pleasure of the company of others."}

\item{six_ils5}{vector for item "There are enough people I feel close to."}

\item{six_ils6}{vector for item "'I often feel rejected."

Items 1, 4 and 6 need to be recoded first}
}
\value{
A numeric vector with sum score of 6 ILS
}
\description{
Calculate sum score of 6 ILS
}
