% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_brs_n}
\alias{calculate_brs_n}
\title{Calculate Brief Resilience Scale (BRS) number of answered questions}
\usage{
calculate_brs_n(brs_1, brs_2, brs_3, brs_4, brs_5, brs_6)
}
\arguments{
\item{brs_1}{vector for item "I tend to bounce back quickly after hard times.."}

\item{brs_2}{vector for item "I have a hard time making it through stressful events."}

\item{brs_3}{vector for item "It does not take me long to recover from a stressful event."}

\item{brs_4}{vector for item "It is hard for me to snap back when something bad happens."}

\item{brs_5}{vector for item "I usually come through difficult times with little trouble."}

\item{brs_6}{vector for item "I tend to take a long time to get over set-backs in my life."}
}
\value{
A numeric vector with number of answered questions of BRS
}
\description{
Count number of anserwed items of BRS
}
