% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_brs_total}
\alias{calculate_brs_total}
\title{Calculate Brief Resilience Scale (BRS) total score}
\usage{
calculate_brs_total(ecu_brs_sum, ecu_brs_n)
}
\arguments{
\item{ecu_brs_sum}{A numeric vector with sum score of BRS}

\item{ecu_brs_n}{A numeric vector with number of answered questions of BRS}
}
\value{
A numeric vector with total score of BRS
}
\description{
Calculate Brief Resilience Scale (BRS) total score
}
