% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{calculate_cfq11_sum}
\alias{calculate_cfq11_sum}
\title{Calculate Chalder Fatigue Scale (CFQ-11) sum score}
\usage{
calculate_cfq11_sum(
  cfq_1,
  cfq_2,
  cfq_3,
  cfq_4,
  cfq_5,
  cfq_6,
  cfq_7,
  cfq_8,
  cfq_9,
  cfq_10,
  cfq_11
)
}
\arguments{
\item{cfq_1}{vector for item "Is fatigue a problem for you?"}

\item{cfq_2}{vector for item "Do you need to rest frequently?."}

\item{cfq_3}{vector for item "Do you feel tired or sleepy?"}

\item{cfq_4}{vector for item "Do you have difficulty getting things done?"}

\item{cfq_5}{vector for item "Do you lack energy?"}

\item{cfq_6}{vector for item "Do you have less strength in your muscles?"}

\item{cfq_7}{vector for item "Do you feel weak?"}

\item{cfq_8}{vector for item "Do you find it difficult to concentrate?"}

\item{cfq_9}{vector for item "Do you have slips of the tongue when you speak?"}

\item{cfq_10}{vector for item "Do you find it difficult to think clearly?"}

\item{cfq_11}{vector for item "How is your memory?"}
}
\value{
A numeric vector with sum score of CFQ-11
}
\description{
Calculate sum score of CFQ-11

items need to be recoded first
}
