% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{calculate_cfs_seid_sum}
\alias{calculate_cfs_seid_sum}
\title{Calculate Chandler Fatigue Scale (CFS/SEID) sum score}
\usage{
calculate_cfs_seid_sum(
  cfs_1,
  cfs_2,
  cfs_3,
  cfs_4,
  cfs_5,
  cfs_6,
  cfs_7,
  cfs_8,
  cfs_9,
  cfs_10,
  cfs_11
)
}
\arguments{
\item{cfs_1}{vector for item "Is fatigue a problem for you?"}

\item{cfs_2}{vector for item "Do you need to rest frequently?."}

\item{cfs_3}{vector for item "Do you feel tired or sleepy?"}

\item{cfs_4}{vector for item "Do you have difficulty getting things done?"}

\item{cfs_5}{vector for item "Do you lack energy?"}

\item{cfs_6}{vector for item "Do you have less strength in your muscles?"}

\item{cfs_7}{vector for item "Do you feel weak?"}

\item{cfs_8}{vector for item "Do you find it difficult to concentrate?"}

\item{cfs_9}{vector for item "Do you have slips of the tongue when you speak?"}

\item{cfs_10}{vector for item "Do you find it difficult to think clearly?"}

\item{cfs_11}{vector for item "How is your memory?"}
}
\value{
A numeric vector with sum score of CFS/SEID
}
\description{
Calculate sum score of CFS/SEID

items need to be recoded first
}
