% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_days_ecu}
\alias{calculate_days_ecu}
\title{Calculate time span in days}
\usage{
calculate_days_ecu(
  start_form = NULL,
  startdate,
  end_form = NULL,
  enddate,
  primary_form = NA
)
}
\arguments{
\item{start_form}{form which includes start date. NULL if one is not working across forms}

\item{startdate}{vector of start dates of time span or column name within start_form containing date variable}

\item{end_form}{form which includes end date. NULL if one is not working across forms}

\item{enddate}{vector of end dates of time span or column name within end_form containing date variable}

\item{primary_form}{"start" if the time span will be merged to start_form. "end" if the time span will be merged to end_form. NA if one is not working across forms}
}
\value{
vector of duration of time span in days
}
\description{
Calculate time span in days
}
\note{
TODO: consider using start_form_mnppid and end_form_mnppid instead of entire forms, see hap/40-core-analysis
}
