% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_dn4_sum}
\alias{calculate_dn4_sum}
\title{Calculate Pain diagnostic questionnaire (DN-4 - Interview) sum score}
\usage{
calculate_dn4_sum(dn4_1, dn4_2, dn4_3, dn4_4, dn4_5, dn4_6, dn4_7)
}
\arguments{
\item{dn4_1}{vector for item "Burning"}

\item{dn4_2}{vector for item "Painful cold"}

\item{dn4_3}{vector for item "Electric Shocks"
Patients were asked "Is the pain associated with one or more of the following symptoms in the same area?"}

\item{dn4_4}{vector for item "Tingling"}

\item{dn4_5}{vector for item "Pins and Needles"}

\item{dn4_6}{vector for item "Numbness"}

\item{dn4_7}{vector for item "Itching"}
}
\value{
A numeric vector with sum score of dn4
}
\description{
Calculate sum score of DN-4

Patients are asked "Does the pain have one or more of the following characteristics?"
}
