% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{calculate_eq5d5l_y_index}
\alias{calculate_eq5d5l_y_index}
\title{Calculate EQ5D-5L-Index (Y Version)}
\usage{
calculate_eq5d5l_y_index(mo, sc, ua, pd, ad)
}
\arguments{
\item{mo}{is a vector for mobility score of eq5d}

\item{sc}{is a vector for self-care score of eq5d}

\item{ua}{is a vector for usual activity score of eq5d}

\item{pd}{is a vector for pain/discomfort score of eq5d}

\item{ad}{is a vector for anxiety/depression score of eq5d}
}
\value{
vector with eq5d-5l-index-y
}
\description{
Calculates EQ-5D-5L index following recommendations for Germany
}
