% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_facitf_scale}
\alias{calculate_facitf_scale}
\title{Calculate Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-Fatigue) subscale score}
\usage{
calculate_facitf_scale(facitf_sum, facitf_n)
}
\arguments{
\item{facitf_sum}{numeric vector with subscale score of FACIT-Fatigue}

\item{facitf_n}{numeric vector with number of answered items of FACIT-Fatigue}
}
\value{
A numeric vector w/ subscale score of FACIT-Fatigue
}
\description{
Calculate FACIT-Fatigue subscale score based on number of answered items
}
