% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_facitf_sum}
\alias{calculate_facitf_sum}
\title{Calculate Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-Fatigue) sum score}
\usage{
calculate_facitf_sum(
  facitf_1,
  facitf_2,
  facitf_3,
  facitf_4,
  facitf_5,
  facitf_6,
  facitf_7,
  facitf_8,
  facitf_9,
  facitf_10,
  facitf_11,
  facitf_12,
  facitf_13,
  facitf_n
)
}
\arguments{
\item{facitf_1}{vector for item "I feel fatigued."}

\item{facitf_2}{vector for item "I feel weak all over."}

\item{facitf_3}{vector for item "I feel listless ("washed out")."}

\item{facitf_4}{vector for item "I feel tired."}

\item{facitf_5}{vector for item "I have trouble starting things because I am tired."}

\item{facitf_6}{vector for item "I have trouble finishing things because I am tired."}

\item{facitf_7}{vector for item "I have energy."}

\item{facitf_8}{vector for item "I am able to do my usual activities."}

\item{facitf_9}{vector for item "I need to sleep during the day."}

\item{facitf_10}{vector for item "I am too tired to eat."}

\item{facitf_11}{vector for item "I need help doing my usual activities."}

\item{facitf_12}{vector for item "I am frustrated by being too tired to do the things I want to do."}

\item{facitf_13}{vector for item "I have to limit my social activity because I am tired."}

\item{facitf_n}{numeric vector with number of answered items of FACIT-Fatigue

Answers were coded /w levels "Not at all", "A little bit", "Somewhat", "Quite a bit" and "Very much"}
}
\value{
A numeric vector with sum score of FACIT-Fatigue
}
\description{
Calculate FACIT-Fatigue sum score

Patients are asked "Please indicate how much each of the following statements has applied to you over the last 7 days by selecting the appropriate item."
}
