% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_fss_sum}
\alias{calculate_fss_sum}
\title{Calculate Fatigue Severity Scale (FSS) sum score}
\usage{
calculate_fss_sum(
  fss_1,
  fss_2,
  fss_3,
  fss_4,
  fss_5,
  fss_6,
  fss_7,
  fss_8,
  fss_9
)
}
\arguments{
\item{fss_1}{vector for item "My motivation is lower when I am fatigued."}

\item{fss_2}{vector for item "Exercise brings on my fatigue."}

\item{fss_3}{vector for item "I am easily fatigued."}

\item{fss_4}{vector for item "Fatigue interferes with my physical functioning."}

\item{fss_5}{vector for item "Fatigue causes frequent problems for me."}

\item{fss_6}{vector for item "My fatigue prevents sustained physical functioning."}

\item{fss_7}{vector for item "Fatigue interferes with carrying out certain duties and responsibilities."}

\item{fss_8}{vector for item "Fatigue is among my three most disabling symptoms."}

\item{fss_9}{vector for item "Fatigue interferes with my work, family or social life."}
}
\value{
A numeric vector with sum score of FSS
}
\description{
Calculate sum score of FSS
}
