% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_fss_sum_phys}
\alias{calculate_fss_sum_phys}
\title{Calculate Fatigue Severity Scale (FSS) sum score for sub component "physical fatigue"}
\usage{
calculate_fss_sum_phys(fss_2, fss_4, fss_6)
}
\arguments{
\item{fss_2}{vector for item "Exercise brings on my fatigue."}

\item{fss_4}{vector for item "Fatigue interferes with my physical functioning."}

\item{fss_6}{vector for item "My fatigue prevents sustained physical functioning."}
}
\value{
A numeric vector with sum score of FSS for sub component "physical fatigue"
}
\description{
Calculate sum score of FSS sub component "physical fatigue"
}
