% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{calculate_pcs_score_hap}
\alias{calculate_pcs_score_hap}
\title{Calculate PCS Score}
\usage{
calculate_pcs_score_hap(ecu_symptom_complex)
}
\arguments{
\item{ecu_symptom_complex}{A dataframe containing symptom complexes}
}
\value{
A dataframe containing PCS-Score
}
\description{
for each symptom-row in long data, apply corresponding the detect_symptom_complex and summarise long symptom data frame by PCSS symptom complex
returns a dataframe with one row per pid containing the results of PCS Score
}
