% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{calculate_pcs_score_suep}
\alias{calculate_pcs_score_suep}
\title{Calculate PCS Score}
\usage{
calculate_pcs_score_suep(
  trial_data,
  pid,
  days_of_pcss_time_diff,
  vector_of_pcss_fup_visits
)
}
\arguments{
\item{trial_data}{SecuTrial Data}

\item{pid}{column name of patient ID in data}

\item{days_of_pcss_time_diff}{Time difference since COVID-19 diagnosis, for which PCSS should be calculated, in days}

\item{vector_of_pcss_fup_visits}{Vector of Follow-Up Visits that are sufficient to evaluate PCSS, based on PCSS time difference}
}
\value{
A dataframe with symptom complexes and PCSS-score for specified Follow-Up visits
}
\description{
for each symptom-row in long data, apply corresponding the detect_symptom_complex and summarise long symptom data frame by PCSS symptom complex
returns a dataframe with one row per pid containing the results of PCS Score
}
