% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_phq15_sum}
\alias{calculate_phq15_sum}
\title{Calculate Patient Health Questionnaire 15-Item Somatic Symptom Severity Scale (PHQ15) sum score}
\usage{
calculate_phq15_sum(
  phq15_1,
  phq15_2,
  phq15_3,
  phq15_4,
  phq15_5,
  phq15_6,
  phq15_7,
  phq15_8,
  phq15_9,
  phq15_10,
  phq15_11,
  phq15_12,
  phq15_13,
  phq15_14,
  phq15_15
)
}
\arguments{
\item{phq15_1}{vector for item "Stomach pain"}

\item{phq15_2}{vector for item "Back pain"}

\item{phq15_3}{vector for item "Pain in your arms, legs, or joints (kneeps, hips, etc.)"}

\item{phq15_4}{vector for item "Menstrual cramps or other problems with your periods [Women only]"}

\item{phq15_5}{vector for item "Headaches"}

\item{phq15_6}{vector for item "Chest pain"}

\item{phq15_7}{vector for item "Dizziness"}

\item{phq15_8}{vector for item "Fainting spells"}

\item{phq15_9}{vector for item "Feeling your heart pound or race"}

\item{phq15_10}{vector for item "Shortness of breath"}

\item{phq15_11}{vector for item "Pain or problems during sexual intercourse"}

\item{phq15_12}{vector for item "Constipation, loos bowels, or diarrhea"}

\item{phq15_13}{vector for item "Nausea, gas, or indigestion"}

\item{phq15_14}{vector for item "Feeling tired or having low energy"}

\item{phq15_15}{vector for item "Trouble sleeping"

Answers were coded /w levels 0 = "Not bothered at all", 1 = "Bothered a little" and 2 = "Bothered a lot"}
}
\value{
A numeric vector with sum score of phq15
}
\description{
Calculate sum score of PHQ15

Patients are asked "During the past 4 weeks, how much have you been bothered by any of the following problems?"
}
