% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_phq4_sum}
\alias{calculate_phq4_sum}
\title{Calculate Patient Health Questionnaire (PHQ 4) sum score}
\usage{
calculate_phq4_sum(phq4_1, phq4_2, phq4_3, phq4_4)
}
\arguments{
\item{phq4_1}{vector for item "Little interest or pleasure in doing things."}

\item{phq4_2}{vector for item "Feeling down, depressed, or hopeless."}

\item{phq4_3}{vector for item "Feeling nervous, anxious or on edge."}

\item{phq4_4}{vector for item "Not being able to stop or control worrying."

Answers need coded levels 0 = "Not at all", 1 = "Several days", 2 = "More than half the days" and 3 = "Nearly every day"}
}
\value{
A numeric vector with sum score of phq8
}
\description{
Calculate sum score of PHQ-4

Patients are asked "Over the last 2 weeks, how often have you been bothered by any of the following problems?"
}
