% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_phq8_sum}
\alias{calculate_phq8_sum}
\title{Calculate Patient Health Questionnaire Depression Scale 8 (PHQ 8) sum score}
\usage{
calculate_phq8_sum(
  phq8_1,
  phq8_2,
  phq8_3,
  phq8_4,
  phq8_5,
  phq8_6,
  phq8_7,
  phq8_8
)
}
\arguments{
\item{phq8_1}{factor for item "Wenig Interesse oder Freude an Ihren Tätigkeiten."}

\item{phq8_2}{factor for item "Niedergeschlagenheit, Schwermut oder Hoffnungslosigkeit."}

\item{phq8_3}{factor for item "Schwierigkeiten, ein- oder durchzuschlafen, oder vermehrter Schlaf."}

\item{phq8_4}{factor for item "Müdigkeit oder Gefühl, keine Energie zu haben."}

\item{phq8_5}{factor for item "Verminderter Appetit oder übermäßiges Bedürfnis zu essen."}

\item{phq8_6}{factor for item "Schlechte Meinung von sich selbst; Gefühl, ein Versager zu sein oder die Familie enttäuscht zu haben."}

\item{phq8_7}{factor for item "Schwierigkeiten, sich auf etwas zu konzentrieren, z. B. beim Zeitung  lesen oder Fernsehen."}

\item{phq8_8}{factor for item "Waren Ihre Bewegungen oder Ihre Sprache so verlangsamt, dass es auch anderen auffallen würde? Oder waren Sie im Gegenteil "zappelig" oder ruhelos und hatten dadurch einen stärkeren Bewegungsdrang als sonst?"}
}
\value{
A numeric vector with sum score of phq8
}
\description{
Calculate sum score of PHQ-8

Patients are asked "Wie oft fühlten Sie sich im Verlauf der letzten 2 Wochen durch die folgenden Beschwerden beeinträchtigt?"
}
