% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_pss_total}
\alias{calculate_pss_total}
\title{Calculate Perceived Stress Scale (PSS) sum score}
\usage{
calculate_pss_total(
  pss1,
  pss2,
  pss3,
  pss4,
  pss5,
  pss6,
  pss7,
  pss8,
  pss9,
  pss10
)
}
\arguments{
\item{pss1}{vector for item "In the last month, how often have you been upset because of something that happened unexpectedly?"}

\item{pss2}{vector for item "In the last month, how often have you felt that you were unable to control the important things in your life?"}

\item{pss3}{vector for item "In the last month, how often have you felt nervous and stressed?"}

\item{pss4}{vector for item "In the last month, how often have you felt confident about your ability to handle your personal problems?"}

\item{pss5}{vector for item "In the last month, how often have you felt that things were going your way?"}

\item{pss6}{vector for item "In the last month, how often have you found that you could not cope with all the things that you had to do?"}

\item{pss7}{vector for item "In the last month, how often have you been able to control irritations in your life?"}

\item{pss8}{vector for item "In the last month, how often have you felt that you were on top of things?"}

\item{pss9}{vector for item "In the last month, how often have you been angered because of things that happened that were outside of your control?"}

\item{pss10}{vector for item "In the last month, how often have you felt difficulties were piling up so high that you could not overcome them?"

Items 4, 5, 7 and 8 need to be recoded first}
}
\value{
A numeric vector with sum score of PSS
}
\description{
Calculate sum score of PSS
}
