% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{calculate_sf36_points3}
\alias{calculate_sf36_points3}
\title{Add points to SF-36 items - 2 groups}
\usage{
calculate_sf36_points3(sf36)
}
\arguments{
\item{sf36}{SF-36 item that gets points}
}
\description{
each answer in SF-36 questionnaire corresponds to a specific amount of points, which differ between items
this function adds points for items with 2 possible answers
}
