% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_6ils_ecu}
\alias{categorize_6ils_ecu}
\title{Categorize Six Item Loneliness Scale (6 ILS)}
\usage{
categorize_6ils_ecu(ecu_6ils_total)
}
\arguments{
\item{ecu_6ils_total}{A numerical vector with 6 ILS sum score}
}
\value{
A factor /w levels "Not lonely" and "Lonely"
}
\description{
Categorize 6 ILS
}
