% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_apache2_score_ecu}
\alias{categorize_apache2_score_ecu}
\title{Categorize Acute Physiology And Chronic Health Evaluation (APACHE) Score}
\usage{
categorize_apache2_score_ecu(apache_score)
}
\arguments{
\item{apache_score}{A numerical vector with APACHE score}
}
\value{
A factorized vector with risk to die during hospitalisation
}
\description{
Categorize APACHE Score in eight categories
}
