% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_barthel_ecu}
\alias{categorize_barthel_ecu}
\title{Categorize barthel index}
\usage{
categorize_barthel_ecu(barthel)
}
\arguments{
\item{barthel}{vector of Barthel index points}
}
\value{
factor of barthel index categories
}
\description{
Categorizes barthel index in four categories
}
