% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_bloodpressure_ecu}
\alias{categorize_bloodpressure_ecu}
\title{Categorize blood pressure}
\usage{
categorize_bloodpressure_ecu(sbp, dbp)
}
\arguments{
\item{sbp}{vector of systolic blood pressure in mmHg}

\item{dbp}{vector of diastolic blood pressure in mmHg}
}
\value{
factor of blood pressure categories
}
\description{
Categorizes blood pressure in seven categories
}
