% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_bmi_ecu}
\alias{categorize_bmi_ecu}
\title{Categorize BMI}
\usage{
categorize_bmi_ecu(bmi, height_uk.factor = NULL, weight_uk.factor = NULL)
}
\arguments{
\item{bmi}{A numeric vector with BMI data. `BMI` unit must be meters per square meter.}

\item{height_uk.factor}{A factor indicating that height is unknown}

\item{weight_uk.factor}{A factor indicating that weight is unknown}
}
\value{
factor of BMI categories
}
\description{
Classify body mass index (BMI) category
bmi_cat(bmi)
}
