% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_brs_ecu}
\alias{categorize_brs_ecu}
\title{Categorize Brief Resilience Scale (BRS)}
\usage{
categorize_brs_ecu(ecu_brs_total)
}
\arguments{
\item{ecu_brs_total}{A numeric vector with total score of BRS}
}
\value{
A factorized vector /w levels "Low resilience", "Normal resilience" and "High resilience"
}
\description{
Categorize Brief Resilience Scale (BRS)
}
