% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{categorize_cfq11}
\alias{categorize_cfq11}
\title{Categorize Chalder Fatigue Scale (CFQ-11)}
\usage{
categorize_cfq11(ecu_cfq11_sum)
}
\arguments{
\item{ecu_cfq11_sum}{A numerical vector with CFQ-11 sum score}
}
\value{
A factor /w levels "No Fatigue" and "Fatigue"
}
\description{
Categorize CFQ-11
}
