% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{categorize_cfs_seid}
\alias{categorize_cfs_seid}
\title{Categorize Chandler Fatigue Scale (CFS/SEID)}
\usage{
categorize_cfs_seid(ecu_cfs_seid_sum)
}
\arguments{
\item{ecu_cfs_seid_sum}{A numerical vector with CFS/SEID sum score}
}
\value{
A factor /w levels "No Fatigue" and "Fatigue"
}
\description{
Categorize CFS/SEID
}
