% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_dds_score_ecu}
\alias{categorize_dds_score_ecu}
\title{Categorize Delirium Detection Score (DDS)}
\usage{
categorize_dds_score_ecu(dds_score)
}
\arguments{
\item{dds_score}{A numerical vector with DDS score}
}
\value{
A factor /w levels "Kein Delir" and "Delir"
}
\description{
Categorize Delirium Detection Score (DDS)
}
