% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_dn4_ecu}
\alias{categorize_dn4_ecu}
\title{Categorize Pain diagnostic questionnaire (DN-4 - Interview)}
\usage{
categorize_dn4_ecu(ecu_dn4_sum)
}
\arguments{
\item{ecu_dn4_sum}{numeric vector with sum score of dn4}
}
\value{
A factorized vector w/ levels "No neuropathic pain" and "Neuropathic pain"
}
\description{
Categorize DN-4 sum score in neuropathic pain and no neuropathic pain
}
