% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_facitf_ecu}
\alias{categorize_facitf_ecu}
\title{Categorize Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-Fatigue)}
\usage{
categorize_facitf_ecu(ecu_facitf_scale)
}
\arguments{
\item{ecu_facitf_scale}{numeric vector with subscale score of FACIT-Fatigue}
}
\value{
A factorized vector w/ levels "Relevant fatigue" and "No fatigue"
}
\description{
Categorize FACIT-Fatigue subscale score in no fatigue and relevant fatigue
}
