% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_fss_mean_ecu}
\alias{categorize_fss_mean_ecu}
\title{Categorize Fatigue Severity Scale (FSS) Mean Score}
\usage{
categorize_fss_mean_ecu(ecu_fss_mean)
}
\arguments{
\item{ecu_fss_mean}{A numerical vector with FSS sum score}
}
\value{
A factor /w levels "No relevant fatigue" and "Relevant fatigue"
}
\description{
Categorize FSS
}
