% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_fss_sum_ecu}
\alias{categorize_fss_sum_ecu}
\title{Categorize Fatigue Severity Scale (FSS) Sum Score}
\usage{
categorize_fss_sum_ecu(ecu_fss_sum)
}
\arguments{
\item{ecu_fss_sum}{A numerical vector with FSS sum score}
}
\value{
A factor /w levels "No to mild fatigue", "Moderate fatigue" and "Severe fatigue"
}
\description{
Categorize FSS
}
