% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_gad7_ecu}
\alias{categorize_gad7_ecu}
\title{Categorize Generalized Anxiety Disorder 7 (GAD-7) sum score}
\usage{
categorize_gad7_ecu(ecu_gad7_sum)
}
\arguments{
\item{ecu_gad7_sum}{numeric vector with sum score of GAD-7}
}
\value{
A factorized vector w/ levels "No anxiety", "Mild anxiety", "Moderate anxiety" and "Severe anxiety"
}
\description{
Categorize GAD-7 sum score in four categories
}
