% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_gcs_ecu}
\alias{categorize_gcs_ecu}
\title{Categorize Glasgow Coma Scale}
\usage{
categorize_gcs_ecu(gcs)
}
\arguments{
\item{gcs}{vector of Glasgow Coma Scale points}
}
\value{
factor of Glasgow Coma Scale categories
}
\description{
Categorizes GCS in four categories
}
