% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_heartfrequency_ecu}
\alias{categorize_heartfrequency_ecu}
\title{Categorize heartfrequency}
\usage{
categorize_heartfrequency_ecu(hf)
}
\arguments{
\item{hf}{vector of heartfrequency in bpm}
}
\description{
returns factor of heartfrequency categories
}
