% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_horowitz_index_ecu}
\alias{categorize_horowitz_index_ecu}
\title{Categorize Horowitz-Index (lung function)}
\usage{
categorize_horowitz_index_ecu(horowitz)
}
\arguments{
\item{horowitz}{A vector with lung function measured in mmHg}
}
\value{
factor of Horowitz categories
}
\description{
Categorize Horowitz-Index in four categories
}
