% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_icdsc_score_ecu}
\alias{categorize_icdsc_score_ecu}
\title{Categorize Intensive Care delir Screening Checklist (ICDSC)}
\usage{
categorize_icdsc_score_ecu(icdsc_score)
}
\arguments{
\item{icdsc_score}{A numerical vector with ICDSC score}
}
\value{
A factor /w levels "No delir", "Subsyndromale delir" and "Delir"
}
\description{
Categorize ICDSC in three categories
}
