% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_mmrc_ecu}
\alias{categorize_mmrc_ecu}
\title{Categorize modified Medical Research Council Dyspnea Scale (MMRC)}
\usage{
categorize_mmrc_ecu(mmrc)
}
\arguments{
\item{mmrc}{factor of modified medical research council dyspnea scale (mmrc) categories}
}
\value{
A factorized vector w/ levels "Dyspnea" and "No disypnea"
}
\description{
Categorize MMRC in dyspnea and no dispnea
}
