% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_moca_ecu}
\alias{categorize_moca_ecu}
\title{Categorize Montreal Cognitive Assessment (MoCA)}
\usage{
categorize_moca_ecu(moca)
}
\arguments{
\item{moca}{vector of MoCA points}
}
\value{
factor of MoCA categories
}
\description{
Categorizes MoCA-Score in two categories
}
