% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_news_score_ecu}
\alias{categorize_news_score_ecu}
\title{Categorize National Early Warning Score (NEWS)}
\usage{
categorize_news_score_ecu(news_score)
}
\arguments{
\item{news_score}{A numerical vector with NEWS score}
}
\value{
A factorized vector w/ levels "Geringes klinisches Risiko", "Moderates Klinisches Risiko", "Hohes Klinisches Risiko", 
"Gering-Moderates klinisches Risiko (Wert 3 in irgendeiner der Kategorien)"
}
\description{
Categorize National Early Warning Score (NEWS)
}
