% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_oxigensaturation_ecu}
\alias{categorize_oxigensaturation_ecu}
\title{Categorize oxigen saturation}
\usage{
categorize_oxigensaturation_ecu(so2)
}
\arguments{
\item{so2}{vector of oxigen saturation SpO2 in percent}
}
\value{
factor of oxigen saturation categories
}
\description{
Categorize oxigen saturation
}
