% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_ph_ecu}
\alias{categorize_ph_ecu}
\title{Categorize pH of blood gas analysis}
\usage{
categorize_ph_ecu(bga_ph)
}
\arguments{
\item{bga_ph}{vector of pH of blood gas analysis}
}
\value{
factor of pH categories
}
\description{
Categorize pH of blood gas analysis
}
