% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_phq15_ecu}
\alias{categorize_phq15_ecu}
\title{Categorize Patient Health Questionnaire 15-Item Somatic Symptom Severity Scale (PHQ15) - 4 groups}
\usage{
categorize_phq15_ecu(ecu_phq15_sum)
}
\arguments{
\item{ecu_phq15_sum}{numeric vector with sum score of phq15}
}
\value{
A factorized vector w/ levels "Minimal somatic symptom severity", "Low somatic symptom severity", "Medium somatic symptom severity", and "High somatic symptom severity
}
\description{
Categorize PHQ-15 sum score in "Minimal somatic symptom severity", "Low somatic symptom severity", "Medium somatic symptom severity", and "High somatic symptom severity"
}
