% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_phq4_ecu}
\alias{categorize_phq4_ecu}
\title{Categorize Patient Health Questionnaire (PHQ 4)}
\usage{
categorize_phq4_ecu(ecu_phq4_sum)
}
\arguments{
\item{ecu_phq4_sum}{numeric vector with sum score of phq4}
}
\value{
A factorized vector w/ levels "No depression", "Mild depression", "Moderate depression" and "Severe depression"
}
\description{
Categorize PHQ-4 sum score in "No depression", "Mild depression", "Moderate depression" and "Severe depression"
}
