% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_phq8_ecu_2}
\alias{categorize_phq8_ecu_2}
\title{Categorize Patient Health Questionnaire Depression Scale 8 (PHQ 8) - 4 groups}
\usage{
categorize_phq8_ecu_2(ecu_phq8_sum)
}
\arguments{
\item{ecu_phq8_sum}{numeric vector with sum score of phq8}
}
\value{
A factorized vector w/ levels "No depression", "Mild depression", "Moderate depression" and "Severe depression"
}
\description{
Categorize PHQ-8 sum score in "No depression", "Mild depression", "Moderate depression" and "Severe depression"
}
