% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_phq9_ecu}
\alias{categorize_phq9_ecu}
\title{Categorize Patient Health Questionnaire Depression Scale 9 (PHQ 9)}
\usage{
categorize_phq9_ecu(ecu_phq9_sum)
}
\arguments{
\item{ecu_phq9_sum}{numeric vector with sum score of phq9}
}
\value{
A factorized vector w/ levels "No depression" and "Depression"
}
\description{
Categorize PHQ-8 sum score in depression and no depression
}
