% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_phq9_ecu_5_cat}
\alias{categorize_phq9_ecu_5_cat}
\title{Categorize Patient Health Questionnaire Depression Scale 9 (PHQ 9) - 5 groups}
\usage{
categorize_phq9_ecu_5_cat(ecu_phq9_sum)
}
\arguments{
\item{ecu_phq9_sum}{numeric vector with sum score of phq9}
}
\value{
A factorized vector w/ levels "Healthy", "No depression", "Mild depression", "Moderate depression" and "Severe depression"
}
\description{
Categorize PHQ-9 sum score in "Healthy", "No depression", "Mild depression", "Moderate depression" and "Severe depression"
}
