% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_phqs_ecu}
\alias{categorize_phqs_ecu}
\title{Categorize Patient Health Questionnaire Stress Scale (PHQ-Stress) - 4 groups}
\usage{
categorize_phqs_ecu(ecu_phqs_sum)
}
\arguments{
\item{ecu_phqs_sum}{numeric vector with sum score of phqs}
}
\value{
A factorized vector w/ levels "Minimal somatic symptom severity", "Low somatic symptom severity", "Medium somatic symptom severity", and "High somatic symptom severity
}
\description{
Categorize PHQ-Stress sum score in "Minimal somatic symptom severity", "Low somatic symptom severity", "Medium somatic symptom severity", and "High somatic symptom severity"
}
