% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_pss_ecu}
\alias{categorize_pss_ecu}
\title{Categorize Perceived Stress Scale (PSS)}
\usage{
categorize_pss_ecu(ecu_pss_total)
}
\arguments{
\item{ecu_pss_total}{A numerical vector with PSS sum score}
}
\value{
A factor /w levels "Low perceived stress", "Moderate perceived stress" and "High perceived stress"
}
\description{
Categorize PSS
}
