% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_resp_rate_ecu}
\alias{categorize_resp_rate_ecu}
\title{Categorize respiration rate}
\usage{
categorize_resp_rate_ecu(resp_rate)
}
\arguments{
\item{resp_rate}{A vector with breaths per minute}
}
\value{
factor of breathing categories
}
\description{
Categorizes respiration rate in three categories
}
