% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_temp_ecu}
\alias{categorize_temp_ecu}
\title{Categorize body temperature}
\usage{
categorize_temp_ecu(temp)
}
\arguments{
\item{temp}{vector of body temperature in degree celsuis}
}
\value{
factor of body temperature categories
}
\description{
Categorize body temperature
}
