% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{create_missing_lookup}
\alias{create_missing_lookup}
\title{Create lookup list for missing values.}
\usage{
create_missing_lookup(
  vallab_lookup,
  missing_labels = NULL,
  missing_values = NULL
)
}
\arguments{
\item{missing_labels}{character vector of value labels that are generally 
considered as missing values.}

\item{missing_values}{numerical vector of values that are generally considered
as missing values.}

\item{lookup}{vallab lookup list.}
}
\value{
nested list.
}
\description{
Create a missing value lookup list by filtering the vallab lookup
list for certain values or value labels that are generally considered as 
missing.
}
\details{
if missing_labels and missing_values are provided to the function,
the missing value lookup list contains entries where either the values matching 
the missing values or the labels matching the missing labels.
}
\keyword{internal}
