% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tsExport.R
\name{dates_tsExport}
\alias{dates_tsExport}
\title{Convert dates in NAPKON Transferstelle Export study data tables}
\usage{
dates_tsExport(tsExport)
}
\arguments{
\item{tsExport}{\code{tsExportdata} that should be dated}
}
\value{
dated \code{tsExportdata}
}
\description{
Convert date variables from string class to Date class.
Keep the original Variables and attach the postfix ".date" to the convertet
variables.

Add date lookup list to export_options
Mark the tsExport object as dated in export_options
}
\seealso{
dates_tsExport_table
}
