% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{defactorize_tsExport}
\alias{defactorize_tsExport}
\title{Remove factorized variables tsExport data.}
\usage{
defactorize_tsExport(tsExport)
}
\arguments{
\item{tsExport}{\code{tsExportdata} where factorized variables should removed.}
}
\value{
\code{tsExportdata} without factor variables found in the vallab 
lookup list.
}
\description{
Remove factor variables that where created during the import
process from data, since they are basically duplicates of the original 
numeric categorical variables and not needed when working with SPSS.
}
\details{
For variables found in the vallab lookup list remove their corresponding 
factorized variables.
Mark the tsExport object as factorized = FALSE in export_options.
}
