% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{detect_symptom_complex}
\alias{detect_symptom_complex}
\title{Calculate symptom complex for Post Covid Syndrome Score (PCSS) calculation}
\usage{
detect_symptom_complex(
  symptom_names,
  vector_of_pcss_fup_visits,
  days_of_pcss_time_diff,
  sy_time_to_visit,
  sy_extracted_name_parent,
  sy_parent,
  symptom_update_visit,
  symptom_update_visit.date,
  sy_end_applicable,
  visit_date.date,
  sy_end_min.date,
  sy_end_max.date,
  sy_dur_min,
  sy_dur_max
)
}
\arguments{
\item{symptom_names}{A vector with names of needed symptoms}

\item{vector_of_pcss_fup_visits}{A vector with labels of desired visit}

\item{days_of_pcss_time_diff}{A vector with number of days of minimum symptom duration}

\item{sy_time_to_visit}{A vector with number of days between symptom start date and visit date}

\item{sy_extracted_name_parent}{A vector with parent form table name for needed symptoms}

\item{sy_parent}{A vector for symptom parent question /w levels "Ja" and "Nein" indicating whether the parent question has been answered with
"yes" or "no"}

\item{symptom_update_visit}{A vector indication the visit at which symptoms were updated last}

\item{symptom_update_visit.date}{A vector indication the date at which symptoms were updated last}

\item{sy_end_applicable}{A vector indicating whether the end date of symptom duration is applicable for time span calculation}

\item{visit_date.date}{A vector with visit date}

\item{sy_end_min.date}{A vector with calculated minimum date of symptom end}

\item{sy_end_max.date}{A vector with calculated maximum date of symptom end}

\item{sy_dur_min}{A vector indicating the minimum number of days of symptom duration}

\item{sy_dur_max}{A vector indicating the maximum number of days of symptom duration}
}
\description{
Calculate symptom complex for Post Covid Syndrome Score (PCSS) calculation
}
