% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{detect_symptom_complex_hap}
\alias{detect_symptom_complex_hap}
\title{Calculate symptom complex for Post-COVID Score (PCS) calculation}
\usage{
detect_symptom_complex_hap(
  ecu_long_symptom_data,
  vector_of_pcss_fup_visit,
  pid,
  trial_data
)
}
\arguments{
\item{ecu_long_symptom_data}{A dataframe containing long_symptom_data}

\item{vector_of_pcss_fup_visit}{A vector with labels of desired visit}

\item{pid}{column name of patient ID in data}

\item{trial_data}{A secuTrial data object}
}
\value{
A dataframe containing symptom complexes per visit in vector_of_pcss_fup_visit
}
\description{
Calculate symptom complex for Post-COVID Score (PCS) calculation
}
