% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{ecu_age_cat_3}
\alias{ecu_age_cat_3}
\title{Categorize age in three intervals}
\usage{
ecu_age_cat_3(age)
}
\arguments{
\item{age}{vector of age in years}
}
\value{
factor of age categories
}
\description{
Categorize age in three intervals
}
