% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{ecu_age_cat_dec}
\alias{ecu_age_cat_dec}
\title{Categorize age in decades}
\usage{
ecu_age_cat_dec(age)
}
\arguments{
\item{age}{vector of age in years}
}
\value{
factor of age categories
}
\description{
Categorize age in decades
}
