% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicodr-package.R
\docType{package}
\name{epicodr-package}
\alias{epicodr}
\alias{epicodr-package}
\title{epicodr: Primary Coding for NAPKON data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Import, primary coding and export functions for SUEP, HAP, POP data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nukleus-ecu/epicodr}
  \item Report bugs at \url{https://github.com/nukleus-ecu/epicodr/issues}
}

}
\author{
\strong{Maintainer}: Klaus Kammerer \email{kammerer_k@ukw.de} (\href{https://orcid.org/0000-0001-8998-0890}{ORCID}) [contributor]

Authors:
\itemize{
  \item Mirjam Kohls \email{kohls_m@ukw.de}
  \item Olga Miljukov \email{miljukov_o@ukw.de}
}

Other contributors:
\itemize{
  \item Institut für Klinische Epidemiologie und Biometrie, Universitätsklinikum Würzburg, Germany \email{ecu@napkon.de} [copyright holder]
}

}
\keyword{internal}
