% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tsExport.R
\name{factorize_tsExport}
\alias{factorize_tsExport}
\title{Factorize in NAPKON Transferstelle Export study data tables}
\usage{
factorize_tsExport(tsExport)
}
\arguments{
\item{tsExport}{\code{tsExportdata} that should get factorized.}
}
\value{
\code{tsExportdata} with value labels
}
\description{
For variables with labels found in the in the cl.csv file.
1. set value labels if not already labelled
2. create factor variables from those variables

Add value_labels lookup list to export_options, if not already present.
Mark the tsExport object as factorized in export_options.
}
\seealso{
factorize_tsExport_table
}
