% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{get_ecu_background_pop}
\alias{get_ecu_background_pop}
\title{Primary codes migration background from nationality and country of birth}
\usage{
get_ecu_background_pop(staatsang, gebland, gebland2)
}
\arguments{
\item{staatsang}{vector "Ist Ihre Staatsangehörigkeit Deutsch?" w/ levels Ja (1); Nein, eine andere (0)}

\item{gebland}{vector "Sind Sie in Deutschland geboren?" w/ levels Ja (1); Nein, in einem anderen Land (0)}

\item{gebland2}{vector "Ist mind. ein Elternteil von Ihnen außerhalb Deutschlands geboren?" w/ levels Ja (1); Nein (0); Weiß nicht (-1)}
}
\value{
A factorized vector w/ levels "Deutschland", 19 different nationalities, or "other"
}
\description{
Primary codes migration background from nationality and country of birth
}
