% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{get_labels_mrs}
\alias{get_labels_mrs}
\title{Label Modified Rankin Scale (mRS)}
\usage{
get_labels_mrs(num_mrs_value)
}
\arguments{
\item{num_mrs_value}{A numerical vector of the mRS}
}
\value{
Label to the numerical values of the mRS
}
\description{
Label mRS
}
