% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{get_labels_mss}
\alias{get_labels_mss}
\title{Label Meningitis Severity Score (MSS)}
\usage{
get_labels_mss(num_mss_value)
}
\arguments{
\item{num_mss_value}{A numerical vector of the MSS score}
}
\value{
Label to the numerical values of the MSS Score
}
\description{
Label MSS
}
