% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{get_time_to_visits}
\alias{get_time_to_visits}
\title{Primary coding time from diagnosis and the onset of symptoms to the selected visit}
\usage{
get_time_to_visits(trial_data, start = "Screening/Baselinevisite")
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{start}{visit which the symptoms and comorbidities are mapped to, default is "Screening/Baselinevisite"}
}
\description{
add the following colums to esym:
inaccuracy, sym_main_date.date, time_diff
}
\details{
add the following columns to ecomorb:
inaccuracy, comorb_oth_diag_d.date, time_diff
}
