% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tsExport.R
\name{label_tsExport}
\alias{label_tsExport}
\title{Set variable labels in NAPKON Transferstelle Export study data tables}
\usage{
label_tsExport(tsExport)
}
\arguments{
\item{tsExport}{\code{tsExportdata} that should be labelled}
}
\value{
labelled \code{tsExportdata}
}
\description{
Set the variable labels and other variable attributes such as unit and 
variable type, for all variables found in the is.csv file.

Add variable lookup list to export_options
Mark the tsExport object as labelled in export_options
}
\seealso{
label_tsExport_table
}
