% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_labels_default}
\alias{missing_labels_default}
\title{Default value labels that are generally considered as missing value.}
\format{
## `Default missing  valuelabels`
Character vector of value labels.
}
\source{
Hand picked labels that appear in the SÜP, HAP and POP study data.
}
\usage{
missing_labels_default
}
\description{
Default value labels that are generally considered as missing value.
}
\keyword{datasets}
