% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_values_default}
\alias{missing_values_default}
\title{Default numeric values generally considered as missing value.}
\format{
## `Default numercial missing values`
}
\source{
numeric value(s) that are standardly used for missing values.
}
\usage{
missing_values_default
}
\description{
Default numeric values generally considered as missing value.
}
\keyword{datasets}
