% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{prepare_spss_export}
\alias{prepare_spss_export}
\title{Prepare SPSS Export}
\usage{
prepare_spss_export(tsExport, rm_stdates = TRUE, set_missings = TRUE, ...)
}
\arguments{
\item{tsExport}{\code{tsExportdata}.}

\item{rm_stdates}{Boolean. If TRUE, original secutrial string date variables
will be deleted, if it has an equivalent date converted variable.}

\item{set_missings}{Boolean. If TRUE, generic missing values will be assigned
to the data.}

\item{...}{Additional arguments passed to \code{prepare_spss_export}.
\itemize{
\item \code{keep_date_vars} defines which original date variables should be kept
when rm_stdates is set to TRUE. Default: "birthdate_score", because it not
really a date and will not represented correctly in SPSS.
\item \code{missing_values} and \code{missing_labels} define variable values and/or
value labels that should be defined as missing values.
Default: \code{missing_labels = missing_labels_default}
\item \code{overwrite_missing_lookup} Boolean Should a new missing value lookup
list created if the tsExportdata object already contains one? Default: TRUE
}}
}
\description{
modify tsExportdata for later export to SPSS files with haven
package used by the write_tsExport function.
}
\details{
The following transformations will be made:
\itemize{
\item assign value labels for variables found in the vallab lookup.
\item remove factor variables if they have an equivalent numeric variable that
is found in the vallab lookup.
\item optional assign general missing values.
\item optional remove original secutrial date string variables that have an
equivalent variable converted to date format.
}
}
\seealso{
\code{\link[=defactorize_tsExport]{defactorize_tsExport()}} \code{\link[=spss_labelled_tsExport]{spss_labelled_tsExport()}}
\code{\link[=remove_st_dates_tsExport]{remove_st_dates_tsExport()}} \code{\link[=spss_missings_tsExport]{spss_missings_tsExport()}}
}
