% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap}
\alias{primary_coding_hap}
\title{Primary coding HAP Data}
\usage{
primary_coding_hap(trial_data)
}
\arguments{
\item{trial_data}{The secu trial data object}
}
\value{
The secu trial data object with primary coded variables
}
\description{
Wrapper function applying the following primary coding steps to trial_data:
}
