% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_age}
\alias{primary_coding_hap_age}
\title{Primary coding age}
\usage{
primary_coding_hap_age(trial_data, table_names)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{table_names}{A vector containing table names of data export}
}
\description{
adds the following columns to demo: 
ecu_age - age in years, ecu_age_cat_dec - age in decades, ecu_age_cat_3 - age in 3 categories
}
