% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_apache2}
\alias{primary_coding_hap_apache2}
\title{Primary coding for Acute Physiology And Chronic Health Evaluation (APACHE) Score}
\usage{
primary_coding_hap_apache2(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to _klinscores1:
ecu_apache2_cat
}
