% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_barthel_pre}
\alias{primary_coding_hap_barthel_pre}
\title{Primary coding Barthel index prior to infection (baseline)}
\usage{
primary_coding_hap_barthel_pre(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to _risiko1:
ecu_barthel_cat_pre
}
