% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_bmi}
\alias{primary_coding_hap_bmi}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_hap_bmi(trial_data, table_names)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{table_names}{A vector containing table names of data export}
}
\description{
adds the following columns to demo: 
ecu_bmi, ecu_bmi_cat
}
