% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_dds}
\alias{primary_coding_hap_dds}
\title{Primary coding for Delirium Detection Score (DDS)}
\usage{
primary_coding_hap_dds(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to _haemodyn:
ecu_dds_cat
}
