% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_eq5d5l}
\alias{primary_coding_hap_eq5d5l}
\title{Primary coding EQ5D-5L-Index}
\usage{
primary_coding_hap_eq5d5l(trial_data, table_names)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{table_names}{A vector containing table names of data export}
}
\description{
adds the following column to eq5d:
ecu_eq5d_index
}
