% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_who_scale}
\alias{primary_coding_hap_who_scale}
\title{Primary coding WHO-Scale}
\usage{
primary_coding_hap_who_scale(trial_data, pid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}
}
\description{
adds the following columns to eosfci
ecu_who_scale.factor, ecu_who_scale
}
\details{
adds the following columns to osfci
ecu_who_scale_max.factor, ecu_who_scale_max
}
