% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop}
\alias{primary_coding_pop}
\title{Primary coding POP Data}
\usage{
primary_coding_pop(trial_data)
}
\arguments{
\item{trial_data}{The secu trial data object}
}
\description{
Wrapper function applying the following primary coding steps to trial_data:
}
