% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_6ils}
\alias{primary_coding_pop_6ils}
\title{Primary coding Six Item Loneliness Scale (6 ILS)}
\usage{
primary_coding_pop_6ils(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to surveyfrageboge: 
ecu_six_ils1,ecu_six_ils2, ecu_six_ils3, ecu_six_ils4, ecu_six_ils5, ecu_six_ils6, ecu_six_ils_total, ecu_six_ils_cat
}
