% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_6mwt}
\alias{primary_coding_pop_6mwt}
\title{6 Minute Walk Test (6MWT)}
\usage{
primary_coding_pop_6mwt(trial_data, pid)
}
\arguments{
\item{trial_data}{a SsecuTrial data object}

\item{pid}{column name of patient ID in trial_data}
}
\description{
adds the following columns to geria:
ecu_6mwt_soll, ecu_6mwt_soll_erreicht
}
