% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_abd_overw}
\alias{primary_coding_pop_abd_overw}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_pop_abd_overw(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to anthropo: 
ecu_waist
}
