% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_age}
\alias{primary_coding_pop_age}
\title{Primary coding age}
\usage{
primary_coding_pop_age(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following columns to erstbefragung: 
ecu_age - age in years, ecu_age_cat_dec - age in decades, ecu_age_cat_3 - age in 3 categories
}
