% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_bmi}
\alias{primary_coding_pop_bmi}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_pop_bmi(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following columns to anthropo: 
ecu_bmi, ecu_bmi_cat, ecu_adipositas
}
