% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_brs}
\alias{primary_coding_pop_brs}
\title{Primary coding Brief Resilience Scale (BRS)}
\usage{
primary_coding_pop_brs(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to surveyfrageboge: 
ecu_brs_sum, ecu_brs_n, ecu_brs_total, ecu_brs_cat
}
