% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_cardio_params}
\alias{primary_coding_pop_cardio_params}
\title{Primary coding cardiological parameters}
\usage{
primary_coding_pop_cardio_params(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variables to kardio: 
ecu_bpm,
}
