% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_clinical_params}
\alias{primary_coding_pop_clinical_params}
\title{Primary coding Clinical Parameters}
\usage{
primary_coding_pop_clinical_params(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following columns to anthropo: 
ecu_temp, ecu_sdp, ecu_dbp, ecu_bp
}
