% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_eq5d5l}
\alias{primary_coding_pop_eq5d5l}
\title{Primary coding EQ5D-5L-Index}
\usage{
primary_coding_pop_eq5d5l(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to erstbefragung: 
ecu_eq5d_index
}
