% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_gpaq_calc}
\alias{primary_coding_pop_gpaq_calc}
\title{Primary coding GPAQ calculation}
\usage{
primary_coding_pop_gpaq_calc(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
Primary coding GPAQ calculation
}
