% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_gpaq_post}
\alias{primary_coding_pop_gpaq_post}
\title{Primary coding Global Physical Activity Questionnaire (GPAQ) post COVID}
\usage{
primary_coding_pop_gpaq_post(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to surveyfrageboge: 
ecu_gpaq_p3_post, ecu_gpaq_p6_post, ecu_gpaq_p9_post, ecu_gpaq_p12_post, ecu_gpaq_p15_post, ecu_gpaq_p16_post, ecu_gpaq_p2cln_post
ecu_gpaq_p3cln_post, ecu_gpaq_p5cln_post, ecu_gpaq_p6cln_post, ecu_gpaq_p8cln_post, ecu_gpaq_p9cln_post, ecu_gpaq_p11cln_post
ecu_gpaq_p12cln_post, ecu_gpaq_p14cln_post, ecu_gpaq_p15cln_post, ecu_gpaq_valid_post, ecu_gpaq_p16cln_post, ecu_gpaq_p1t3cln_post 
ecu_gpaq_p4t6cln_post, ecu_gpaq_p7t9cln_post, ecu_gpaq_p10t12cln_post, ecu_gpaq_p13t15cln_post, ecu_gpaq_p1t3_met_post, ecu_gpaq_p4t6_met_post
ecu_gpaq_p7t9_met_post, ecu_gpaq_p10t12_met_post, ecu_gpaq_p13t15_me_post, ecu_gpaq_ptotal_met_post, ecu_gpaq_ptotalday_met_post
ecu_gpaq_pworkday_met_post, ecu_gpaq_ptravelday_met_post, ecu_gpaq_precday_met_post, ecu_gpaq_per_work_met_post, ecu_gpaq_per_trans_met_post
ecu_gpaq_per_rec_met_post, ecu_gpaq_met_who_post, ecu_gpaq_p1t3_uw_post, ecu_gpaq_p4t6_uw_post, ecu_gpaq_p7t9_uw_post, ecu_gpaq_p10t12_uw_post
ecu_gpaq_p13t15_uw_post, ecu_gpaq_ptotal_uw_post, ecu_gpaq_ptotalday_uw_post, ecu_gpaq_pworkday_uw_post, ecu_gpaq_ptravelday_uw_post
ecu_gpaq_precday_uw_post, ecu_gpaq_per_work_uw_post, ecu_gpaq_per_trans_uw_post, ecu_gpaq_per_rec_uw_post, ecu_gpaq_work_post, ecu_gpaq_trans_post 
ecu_gpaq_rec_post, ecu_gpaq_vig_activ_post, ecu_gpaq_cln_post, ecu_gpaq_cln_sedentary_post
}
