% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_gpaq_pre}
\alias{primary_coding_pop_gpaq_pre}
\title{Primary coding Global Physical Activity Questionnaire (GPAQ) pre COVID}
\usage{
primary_coding_pop_gpaq_pre(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to surveyfrageboge: 
ecu_gpaq_p3, ecu_gpaq_p6, ecu_gpaq_p9, ecu_gpaq_p12, ecu_gpaq_p15, ecu_gpaq_p16, ecu_gpaq_p2cln, ecu_gpaq_p3cln
ecu_gpaq_p5cln, ecu_gpaq_p6cln, ecu_gpaq_p8cln, ecu_gpaq_p9cln, ecu_gpaq_p11cln, ecu_gpaq_p12cln, ecu_gpaq_p14cln
ecu_gpaq_p15cln, ecu_gpaq_valid, ecu_gpaq_p16cln, ecu_gpaq_p1t3cln, ecu_gpaq_p4t6cln, ecu_gpaq_p7t9cln, ecu_gpaq_p10t12cln
ecu_gpaq_p13t15cln, ecu_gpaq_p1t3_met, ecu_gpaq_p4t6_met, ecu_gpaq_p7t9_met, ecu_gpaq_p10t12_met, ecu_gpaq_p13t15_met
ecu_gpaq_ptotal_met, ecu_gpaq_ptotalday_met, ecu_gpaq_pworkday_met, ecu_gpaq_ptravelday_met, ecu_gpaq_precday_met
ecu_gpaq_per_work_met, ecu_gpaq_per_trans_met, ecu_gpaq_per_rec_met, ecu_gpaq_met_who, ecu_gpaq_p1t3_uw, ecu_gpaq_p4t6_uw
ecu_gpaq_p7t9_uw, ecu_gpaq_p10t12_uw, ecu_gpaq_p13t15_uw, ecu_gpaq_ptotal_uw, ecu_gpaq_ptotalday_uw, ecu_gpaq_pworkday_uw
ecu_gpaq_ptravelday_uw, ecu_gpaq_precday_uw, ecu_gpaq_per_work_uw, ecu_gpaq_per_trans_uw, ecu_gpaq_per_rec_uw
ecu_gpaq_work, ecu_gpaq_trans, ecu_gpaq_rec, ecu_gpaq_vig_activ, ecu_gpaq_cln, ecu_gpaq_cln_sedentary
}
