% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_gpaq_vars_post}
\alias{primary_coding_pop_gpaq_vars_post}
\title{Primary coding GPAQ variable/column selection for GPAQ post COVID}
\usage{
primary_coding_pop_gpaq_vars_post(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
Primary coding GPAQ variable/column selection for GPAQ post COVID
}
