% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_kccq}
\alias{primary_coding_pop_kccq}
\title{Kansas City Cardiomyopathy Questionnaire (KCCQ)}
\usage{
primary_coding_pop_kccq(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
calculates Kansas City Cardiomyopathy Questionnaire (KCCQ) total score
}
\details{
adds the following variables to surveyfrageboge:
ecu_kccq_phys_sum, ecu_kccq_phys_score, ecu_kccq_sy_ch, ecu_kccq_sy_ch_score, ecu_kccq_3_score,
ecu_kccq_5_score, ecu_kccq_7_score, ecu_kccq_9_score, ecu_kccy_sy_freq_score, ecu_kccq_sy_sev_sum,
ecu_kccq_sy_sev_score, ecu_kccq_sy_sev_sum, ecu_kccq_sy_sev_score, ecu_kccq_sy_sum, ecu_kccq_sy_score,
ecu_kccq_se_sum, ecu_kccq_se_score, ecu_kccq_qol_sum, ecu_kccq_qol_score, ecu_kccq_sl_sum,
ecu_kccq_sl_score, ecu_kccq_total
}
