% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_migration}
\alias{primary_coding_pop_migration}
\title{Primary migration background}
\usage{
primary_coding_pop_migration(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to erstbefragung: 
ecu_migration_background
}
