% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_mmrc}
\alias{primary_coding_pop_mmrc}
\title{Primary coding modified Medical Research Council Dyspnea Scale (mMRC)}
\usage{
primary_coding_pop_mmrc(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to surveyfragebogen: 
ecu_mmrc
}
