% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_moca}
\alias{primary_coding_pop_moca}
\title{Primary coding Montreal Cognitive Assessment (MoCA)}
\usage{
primary_coding_pop_moca(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to neuro
ecu_moca_total_score, ecu_moca_cat
}
