% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_phq8}
\alias{primary_coding_pop_phq8}
\title{Primary coding Patient Health Questionnaire depession scale (PHQ-8)}
\usage{
primary_coding_pop_phq8(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to surveyfrageboge: 
ecu_phq8
}
