% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_pneumo_params}
\alias{primary_coding_pop_pneumo_params}
\title{Primary coding cardiological parameters}
\usage{
primary_coding_pop_pneumo_params(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to kardio: 
ecu_spo2,
}
