% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_psqi}
\alias{primary_coding_pop_psqi}
\title{Pittsburgh Sleep Quality Index}
\usage{
primary_coding_pop_psqi(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
add the following columns to surveyfrageboge:
ecu_psqi_comp_1, ecu_psqi_comp_2_sum, ecu_psqi_comp_2, ecu_psqi_comp_3, ecu_psqi_comp_4, ecu_psqi_comp_5_sum, ecu_psqi_comp_5, ecu_psqi_comp_6, 
ecu_psqi_comp_7_sum, ecu_psqi_comp_7, ecu_psqi_global_score
}
