% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_pss}
\alias{primary_coding_pop_pss}
\title{Primary coding Perceived Stress Scale (PSS)}
\usage{
primary_coding_pop_pss(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to surveyfrageboge: 
ecu_pss1,ecu_pss2, ecu_pss3, ecu_pss4, ecu_pss5, ecu_pss6, ecu_pss_total, ecu_pss_cat
}
