% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive}
\alias{primary_coding_rapid_revive}
\title{Primary coding RAPID REVIVE Data}
\usage{
primary_coding_rapid_revive(trial_data)
}
\arguments{
\item{trial_data}{The secu trial data object}
}
\description{
Wrapper function applying the following primary coding steps to trial_data:
}
