% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_age}
\alias{primary_coding_rapid_revive_age}
\title{Primary coding age}
\usage{
primary_coding_rapid_revive_age(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to demo: 
ecu_age_cat_dec - age in decades, ecu_age_cat_3 - age in 3 categories
}
