% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_bmi}
\alias{primary_coding_rapid_revive_bmi}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_rapid_revive_bmi(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to demo: 
ecu_bmi, ecu_bmi_cat, ecu_adipositas
}
