% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_bp}
\alias{primary_coding_rapid_revive_bp}
\title{Primary coding blood pressure}
\usage{
primary_coding_rapid_revive_bp(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to vital: 
ecu_bp
}
