% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_fss}
\alias{primary_coding_rapid_revive_fss}
\title{Primary coding Fatigue Severity Scale (FSS)}
\usage{
primary_coding_rapid_revive_fss(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following colums to fss
}
