% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_gad7}
\alias{primary_coding_rapid_revive_gad7}
\title{Primary coding Generalized Anxiety Disorder 7 (GAD-7)}
\usage{
primary_coding_rapid_revive_gad7(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to gad7:
ecu_gad7_sum
ecu_gad_cat
}
