% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_hf}
\alias{primary_coding_rapid_revive_hf}
\title{Primary coding heart frequency}
\usage{
primary_coding_rapid_revive_hf(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to vital: 
ecu_hr
}
