% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_lab}
\alias{primary_coding_rapid_revive_lab}
\title{Primary coding laboratory units}
\usage{
primary_coding_rapid_revive_lab(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
add the following colums to labor:
ecu_lab_leuko_pnl, ecu_lab_basoph_pnl, ecu_lab_eoph_pnl, ecu_lab_lympho_pnl, ecu_lab_haema_perc, ecu_lab_haemog_gpdl, ecu_lab_eryt_ppl, ecu_lab_mch_pg, ecu_lab_mchc_gpdl, 
ecu_lab_mono_pnl, ecu_lab_neutroph_pnl, ecu_lab_thromb_pnl, ecu_lab_crea_mgpdl, ecu_lab_bun_mgpdl, ecu_lab_uricacid_mgpdl, ecu_lab_ast_Upl, ecu_lab_alt_Upl, ecu_lab_ap_Upl, 
ecu_lab_ygt_Upl, ecu_lab_bilitotal_mgpdl, ecu_lab_bilidir_mgpdl, ecu_lab_amy_Upl, ecu_lab_lipase_Upl, ecu_lab_glc_mgpdl, ecu_lab_glcmin_mgpdl, ecu_lab_glcmax_mgpdl, ecu_lab_trigly_mgpdl
}
