% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_mmrc}
\alias{primary_coding_rapid_revive_mmrc}
\title{Primary coding modified Medical Research Council Dyspnea Scale (mMRC)}
\usage{
primary_coding_rapid_revive_mmrc(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to mmrc: 
ecu_mmrc
}
