% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_moca}
\alias{primary_coding_rapid_revive_moca}
\title{Primary cpding for Montreal Cognitive Assessment (MoCA)}
\usage{
primary_coding_rapid_revive_moca(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following varible to moca
ecu_moca_total_score, ecu_moca_cat
}
