% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_pem}
\alias{primary_coding_rapid_revive_pem}
\title{Primary coding Post Exertional Malaise (PEM) questionnaire}
\usage{
primary_coding_rapid_revive_pem(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to pem: 
ecu_pem, ecu_mecfs
}
