% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_phq15}
\alias{primary_coding_rapid_revive_phq15}
\title{Primary coding Patient Health Questionnaire 15-Item Somatic Symptom Severity Scale (PHQ15)}
\usage{
primary_coding_rapid_revive_phq15(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to phq15: 
ecu_phq15_sum, ecu_phq15_cat
}
