% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_phq9}
\alias{primary_coding_rapid_revive_phq9}
\title{Primary coding Patient Health Questionnaire Depession Scale (PHQ-9)}
\usage{
primary_coding_rapid_revive_phq9(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to phq9: 
ecu_phq9_sum, ecu_phq9_binary, ecu_phq9_5_cat
}
