% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_phqs}
\alias{primary_coding_rapid_revive_phqs}
\title{Primary coding Patient Health Questionnaire Stress Scale (PHQ-Stress)}
\usage{
primary_coding_rapid_revive_phqs(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to phqs: 
ecu_phqs_sum, ecu_phqs_cat
}
