% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_resp_rate}
\alias{primary_coding_rapid_revive_resp_rate}
\title{Primary coding respiration rate}
\usage{
primary_coding_rapid_revive_resp_rate(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to eohrsub: 
ecu_os_resp_rate
}
