% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid}
\alias{primary_coding_snid}
\title{Primary coding SNID Data}
\usage{
primary_coding_snid(trial_data)
}
\arguments{
\item{trial_data}{The secu trial data object}
}
\description{
Wrapper function applying the following primary coding steps to trial_data:
}
