% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_age}
\alias{primary_coding_snid_age}
\title{Primary coding age}
\usage{
primary_coding_snid_age(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to patinf: 
ecu_age_cat_dec - age in decades, ecu_age_cat_3 - age in 3 categories
}
