% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_bmi}
\alias{primary_coding_snid_bmi}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_snid_bmi(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to patinf: 
ecu_bmi, ecu_bmi_cat, ecu_adipositas
}
