% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_gcs}
\alias{primary_coding_snid_gcs}
\title{Primary coding Glasgow Coma Scale}
\usage{
primary_coding_snid_gcs(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to vital: 
ecu_gcs
}
