% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_lab}
\alias{primary_coding_snid_lab}
\title{Primary coding laboratory units}
\usage{
primary_coding_snid_lab(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
add the following colums to lab:
ecu_lab_leuko_1E9pliter, ecu_lab_neutroph_1E9pliter, ecu_lab_lympho_1E9pliter, ecu_lab_crp_mgpliter,  
ecu_lab_pct_ngpml, ecu_lab_il6_pgpml, ecu_lab_ast_Upl, ecu_lab_crea_mgpdl, ecu_lab_lactate_mmolpl
}
