% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_mms}
\alias{primary_coding_snid_mms}
\title{Primary coding Meningitis Severity Score (MSS)}
\usage{
primary_coding_snid_mms(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to scorecns: 
ecu_mss_label
}
