% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_mrs}
\alias{primary_coding_snid_mrs}
\title{Primary Coding Modified Rankin Scale}
\usage{
primary_coding_snid_mrs(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to scorecns:
ecu_mrs_label
}
