% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_oxy_sat}
\alias{primary_coding_snid_oxy_sat}
\title{Primary coding respiration rate}
\usage{
primary_coding_snid_oxy_sat(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to vital: 
ecu_oxy_sat
}
