% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-snid.R
\name{primary_coding_snid_temp}
\alias{primary_coding_snid_temp}
\title{Primary coding temperature}
\usage{
primary_coding_snid_temp(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to vital: 
ecu_temp
}
