% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep}
\alias{primary_coding_suep}
\title{Primary coding SUEP Data}
\usage{
primary_coding_suep(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\value{
A secuTrial data object with primary coded variables and dataframes
}
\description{
Wrapper function applying the following primary coding steps to trial_data: 
primary_coding_suep_age, primary_coding_suep_bmi, primary_coding_suep_clinical_params, 
primary_coding_suep_baseline_barthel, primary_coding_suep_patient_barthel, primary_coding_suep_moca
}
