% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_age}
\alias{primary_coding_suep_age}
\title{Primary coding age}
\usage{
primary_coding_suep_age(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{SecuTrial Data}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to bv1: ecu_age - age in years, ecu_age_cat_dec - age in decades, ecu_age_cat_3 - age in 3 categories
}
