% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_ards}
\alias{primary_coding_suep_ards}
\title{Primary coding for ARDS by imaging procedures}
\usage{
primary_coding_suep_ards(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\value{
A secuTrial data object with primary coded variables and dataframes
}
\description{
adds the following dataframes to data: 
ecu_ards
}
