% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_baseline_barthel}
\alias{primary_coding_suep_baseline_barthel}
\title{Primary coding Barthel index prior to infection (baseline)}
\usage{
primary_coding_suep_baseline_barthel(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{SecuTrial Data}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following variable to bv1:
ecu_bl_barthel_cat
}
\details{
uses helper function categorize_barthel_ecu
}
