% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_bmi}
\alias{primary_coding_suep_bmi}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_suep_bmi(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{SecuTrial Data}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to fuv1: ecu_bmi, ecu_bmi_est, ecu_bmi_cat, ecu_bmi_est_cat
}
