% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_cfq11}
\alias{primary_coding_suep_cfq11}
\title{Primary coding Chalder Fatigue Scale (CFQ-11)}
\usage{
primary_coding_suep_cfq11(trial_data, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to promext:
ecu_cfq11_1,ecu_cfq11_2, ecu_cfq11_3, ecu_cfq11_4, ecu_cfq11_5, ecu_cfq11_6, ecu_cfq11_7, ecu_cfq11_8, ecu_cfq11_9,
ecu_cfq11_10, ecu_cfq11_11, ecu_cfq11_sum, ecu_cfq11_cat
}
