% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_cfs_seid}
\alias{primary_coding_suep_cfs_seid}
\title{Primary coding Chandler Fatigue Scale (CFS/SEID)}
\usage{
primary_coding_suep_cfs_seid(trial_data, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to promext:
ecu_cfs_seid_1,ecu_cfs_seid_2, ecu_cfs_seid_3, ecu_cfs_seid_4, ecu_cfs_seid_5, ecu_cfs_seid_6, ecu_cfs_seid_7, ecu_cfs_seid_8, ecu_cfs_seid_9,
ecu_cfs_seid_10, ecu_cfs_seid_11, ecu_cfs_seid_sum, ecu_cfs_seid_cat
}
