% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_dn4}
\alias{primary_coding_suep_dn4}
\title{Primary coding Pain diagnostic Questionnaire (DN-4 - Interview)}
\usage{
primary_coding_suep_dn4(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to promext: 
ecu_dn4_sum, ecu_dn4_cat
}
