% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_eq5d3l_13t18p}
\alias{primary_coding_suep_eq5d3l_13t18p}
\title{Primary coding EQ5D-3L-Index for children aged 13-18 (PROXY Version)}
\usage{
primary_coding_suep_eq5d3l_13t18p(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following column to promp13:
ecu_eq5d3l_proxy_index
}
