% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_eq5d5l}
\alias{primary_coding_suep_eq5d5l}
\title{Primary coding EQ5D-5L-Index}
\usage{
primary_coding_suep_eq5d5l(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following variable to prom:
ecu_eq5d_index
}
