% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_moca}
\alias{primary_coding_suep_moca}
\title{Primary coding MoCA}
\usage{
primary_coding_suep_moca(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{SecuTrial Data}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following variable to fuv3:
ecu_moca_cat
}
