% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_patient_barthel}
\alias{primary_coding_suep_patient_barthel}
\title{Primary coding Barthel index (patient status)}
\usage{
primary_coding_suep_patient_barthel(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following variable to fuv1:
ecu_patient_barthel_cat
}
