% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_pcs_score}
\alias{primary_coding_suep_pcs_score}
\title{Primary coding PCS-Score}
\usage{
primary_coding_suep_pcs_score(trial_data, prom = "No")
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{prom}{A vector indication whether proms should be included in pcs calculation or not; needs to be specified as prom = "No" or prom = "Yes"}
}
\value{
A secuTrial data object with primary coded variables and dataframes
}
\description{
adds the following dataframes to data: 
ecu_long_symptom_data, ecu_pcs_score
}
