% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_phq4}
\alias{primary_coding_suep_phq4}
\title{Primary coding Patient Health Questionnaire (PHQ-4)}
\usage{
primary_coding_suep_phq4(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to prom: 
ecu_phq4_1, ecu_phq4_2, ecu_phq4_3, ecu_phq4_4, ecu_phq4_sum, ecu_phq4_cat
}
