% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_recode_brs}
\alias{primary_coding_suep_recode_brs}
\title{Primary coding Recode Brief Resilience Scale (BRS)}
\usage{
primary_coding_suep_recode_brs(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
recodes the following columns in promext: 
brs_2, brs_4, brs_6
}
