% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_recode_promis_cogn_funct}
\alias{primary_coding_suep_recode_promis_cogn_funct}
\title{Primary coding PROMIS-29 Cognitive impairments items (recoding)}
\usage{
primary_coding_suep_recode_promis_cogn_funct(trial_data, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to prom:
ecu_pro_cogn_1, ecu_pro_cogn_2, ecu_pro_cogn_3, ecu_pro_cogn_4
}
