% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_recode_promis_dysp}
\alias{primary_coding_suep_recode_promis_dysp}
\title{Primary coding PROMIS-29 Dyspnoe items (recoding)
 
adds the following columns to promext: 
ecu_pro_dysp_1, ecu_pro_dysp_2, ecu_pro_dysp_3, ecu_pro_dysp_4, ecu_pro_dysp_5, ecu_pro_dysp_6, ecu_pro_dysp_7,
ecu_pro_dysp_8, ecu_pro_dysp_9, ecu_pro_dysp_10}
\usage{
primary_coding_suep_recode_promis_dysp(trial_data, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{visitid}{column name of visit ID in trial_data}
}
\value{
A secuTrial data object with primary coded variables and dataframes
}
\description{
Primary coding PROMIS-29 Dyspnoe items (recoding)
 
adds the following columns to promext: 
ecu_pro_dysp_1, ecu_pro_dysp_2, ecu_pro_dysp_3, ecu_pro_dysp_4, ecu_pro_dysp_5, ecu_pro_dysp_6, ecu_pro_dysp_7,
ecu_pro_dysp_8, ecu_pro_dysp_9, ecu_pro_dysp_10
}
