% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_recode_promis_phys}
\alias{primary_coding_suep_recode_promis_phys}
\title{Primary coding PROMIS-29 Physical Function items (recoding)}
\usage{
primary_coding_suep_recode_promis_phys(trial_data, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to promext:
ecu_pro29_phys_funct_1, ecu_pro29_phys_funct_2, ecu_pro29_phys_funct_3, ecu_pro29_phys_funct_4
}
