% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_recode_promis_sleep}
\alias{primary_coding_suep_recode_promis_sleep}
\title{Primary coding PROMIS-29 Sleep disturbance items (recoding)}
\usage{
primary_coding_suep_recode_promis_sleep(trial_data, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to promext:
ecu_pro29_sleep_1, ecu_pro29_sleep_2
}
