% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_recode_promis_social}
\alias{primary_coding_suep_recode_promis_social}
\title{Primary coding PROMIS-29 Ability to participate in social roles and activities items (recoding)}
\usage{
primary_coding_suep_recode_promis_social(trial_data, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following columns to promext:
ecu_pro29_social_1, ecu_pro29_social_2, ecu_pro29_social_3, ecu_pro29_social_4
}
