% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_who_scale}
\alias{primary_coding_suep_who_scale}
\title{Primary coding WHO-Scale}
\usage{
primary_coding_suep_who_scale(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}

\item{visitid}{column name of visit ID in trial_data}
}
\description{
adds the following dataframe to trial_data: 
 ecu_who_scale_per_visit_data
 
adds the following columns to scv:
 ecu_who_scale_max, ecu_who_scale_max.factor, ecu_who_scale_with_diag_max, ecu_who_scale_with_diag_max.factor
}
