% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tsExport.R
\name{read_tsExport}
\alias{read_tsExport}
\title{Read NAPKON Transferstelle export}
\usage{
read_tsExport(
  data_dir,
  separator = ";",
  decimal = ".",
  encoding = "UTF-8",
  quote_escape_detect = TRUE,
  escape_backslash = TRUE,
  labels = TRUE,
  factor = TRUE,
  dates = TRUE
)
}
\arguments{
\item{data_dir}{string - location of the export}

\item{separator}{Character. CSV file column separator.}

\item{decimal}{Character.}

\item{encoding}{Character. Encoding}

\item{quote_escape_detect}{Boolean.Should the function try to detect 
the escaping method used in the data file? If TRUE, the reading process 
might be slower.}

\item{escape_backslash}{Boolean.Are Quotes in strings escaped by a
backslash? (The default with (german) secuTrial Exports).}

\item{labels}{logical - add labels to variables and table}

\item{factor}{logical - convert categorical variables to factor variables
(ignored when reference values are not in a separate table)}

\item{dates}{logical - convert date variables}
}
\value{
\code{tsExportdata} object - a list with one data.frame for each file 
  on the export and a list containing the export options
}
\description{
Convenience wrapper for \code{read_tsExport_raw}, ' \code{label_tsExport},
\code{factorize_tsExport} and \code{dates_tsExport}.
}
