% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tsExport.R
\name{read_tsExport_raw}
\alias{read_tsExport_raw}
\title{Load NAPKON Transferstelle export data from directory or zip file.}
\usage{
read_tsExport_raw(
  data_dir,
  separator = ";",
  decimal = ",",
  encoding = "UTF-8",
  quote_escape_detect = TRUE,
  escape_backslash = TRUE
)
}
\arguments{
\item{data_dir}{String. Path with directory or zip file of the Transferstelle 
export.}

\item{separator}{Character. CSV file column separator.}

\item{decimal}{Character.}

\item{encoding}{Character. Encoding}

\item{quote_escape_detect}{Boolean.Should the function try to detect 
the escaping method used in the data file? If TRUE, the reading process 
might be slower.}

\item{escape_backslash}{Boolean.Are Quotes in strings escaped by a
backslash? (The default with (german) secuTrial Exports).}
}
\value{
\code{tsExportdata} Object containing a list of export data 
  information and data.frames with all the data loaded from the 
  Transferstelle export.
}
\description{
Load the full set of meta and data tables. 
Generate and store export information in export_options.
}
