% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_brs}
\alias{recode_brs}
\title{Recode Brief Resilience Scale items}
\usage{
recode_brs(brs.factor)
}
\arguments{
\item{brs}{BRS item, that needs recoding}
}
\value{
A numerical vector with recoded BRS item
}
\description{
item polarization varies across BRS items 
item 1, 3, 5 need levels from 1 = "Do not agree at all" to 5 = "Totally agree" 
item 2, 4, 6 need levels from 5 = "Do not agree at all" to 1 = "Totally agree"
function recodes items 2, 4, 6 to needed levels if they were leveled like item 1, 3 and 5
}
