% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_cfq11}
\alias{recode_cfq11}
\title{Recode Chalder Fatigue Scale (CFQ-11) items}
\usage{
recode_cfq11(cfq)
}
\arguments{
\item{cfq}{Item of Chalder Fatigue Scale (CFQ-11) that needs to be recoded}
}
\value{
A vector with the correctly coded CFQ-11 item
}
\description{
recodes the Chalder Fatigue Scale (CFQ-11) items into a dichotomous variable
}
