% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_cfs_seid}
\alias{recode_cfs_seid}
\title{Recode Chandler Fatigue Scale (CFS/SEID) items}
\usage{
recode_cfs_seid(cfs)
}
\arguments{
\item{cfs}{Item of Chandler Fatigue Scale (CFS/SEID) that needs to be recoded}
}
\value{
A vector with the correctly coded CFS/SEID item
}
\description{
recodes the Chandler Fatigue Scale (CFS/SEID) items into a dichotomous variable
}
