% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{recode_dn4}
\alias{recode_dn4}
\title{Recode Pain diagnostic questionnaire (DN-4 - Interview)}
\usage{
recode_dn4(dn4)
}
\arguments{
\item{dn4}{DN-4 item, that needs recoding}
}
\description{
recodes Pain diacnostic questionnaire items to needed levels
}
