% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{recode_phq}
\alias{recode_phq}
\title{Recode Patient Health Questionnaire items (PHQ)}
\usage{
recode_phq(phq.factor)
}
\arguments{
\item{phq.factor}{PHQ item, that needs recoding}
}
\description{
recodes Patient Health Questionnaire items to needed levels
}
