% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_promis29_phys_funct}
\alias{recode_promis29_phys_funct}
\title{Recode PROMIS-29 Physical Function items}
\usage{
recode_promis29_phys_funct(pro_phys_funct.factor)
}
\arguments{
\item{pro_phys_funct.factor}{Item of PROMIS-29 Physical Function that needs to be recoded}
}
\value{
A vector with the correctly coded PROMIS-29 Physical Function item
}
\description{
recodes the PROMIS-29 Physical Function items to needed levels
}
