% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_promis29_sleep}
\alias{recode_promis29_sleep}
\title{Recode PROMIS-29 Sleep disturbance items}
\usage{
recode_promis29_sleep(pro_sleep_var.factor)
}
\arguments{
\item{pro_sleep_var.factor}{Item of PROMIS-29 Sleep disturbance that needs to be recoded}
}
\value{
A vector with the correctly coded PROMIS-29 Sleep disturbance item
}
\description{
recodes the PROMIS-29 Sleep disturbance items to needed levels
}
