% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_promis29_social}
\alias{recode_promis29_social}
\title{Recode PROMIS-29 Ability to participate in social roles and activities items}
\usage{
recode_promis29_social(pro_social.factor)
}
\arguments{
\item{pro_social.factor}{Item of PROMIS-29 Ability to participate in social roles and activities that needs to be recoded}
}
\value{
A vector with the correctly coded PROMIS-29 Ability to participate in social roles and activities item
}
\description{
recodes the PROMIS-29 Ability to participate in social roles and activities items to needed levels
}
