% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_promis_cognitive}
\alias{recode_promis_cognitive}
\title{Recode PROMIS-29 Cognitive impairments items}
\usage{
recode_promis_cognitive(pro_cogn_var.factor)
}
\arguments{
\item{pro_cogn_var.factor}{Item of PROMIS-29 Cognitive impairments that needs to be recoded}
}
\value{
A vector with the correctly coded PROMIS-29 Cognitive impairments item
}
\description{
recodes the PROMIS-29 Cognitive impairments items to needed levels
}
