% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{recode_promis_dyspnoe}
\alias{recode_promis_dyspnoe}
\title{Recode PROMIS-29 Dyspnea items}
\usage{
recode_promis_dyspnoe(pro_dysp_var.factor)
}
\arguments{
\item{pro_dysp_var.factor}{Item of PROMIS-29 Dyspnea that needs to be recoded}
}
\value{
A vector with the correctly coded PROMIS-29 Dyspnea item
}
\description{
recodes the PROMIS-29 Dyspnea items to needed levels
}
