% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{recode_pss}
\alias{recode_pss}
\title{Recode Perceived Stress Scale (PSS) items}
\usage{
recode_pss(pss.factor, version)
}
\arguments{
\item{pss.factor}{A factor containing the factorized PSS item that needs recoding}

\item{version}{a vector indication whether item wording was positive or negative}
}
\description{
recodes the PSS items depending on positive or negative wording
item 1, 2, 3, 6, 9 and 10 need levels from 0 = "Never" to 4 = "Very often" 
item 2, 4, 7 and 8 need levels from 4 = "Never" to 0 = "Very often"
function recodes items to needed levels
}
