% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{recode_sf36}
\alias{recode_sf36}
\title{Recode Short Form Survey Instrument (SF-36) items}
\usage{
recode_sf36(sf36_item.factor)
}
\arguments{
\item{sf36_item.factor}{SF-36 item that needs recoding}
}
\description{
item coding is wrong in some items of SF-36 questionnaire
current levels: 1 = "Yes" and 0  = "No"
needed levels: 1 = "Yes" and 2 = "No" 
function recodes items to needed levels if they were coded incorrectly
}
