% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{remove_st_dates_tsExport}
\alias{remove_st_dates_tsExport}
\title{Remove original secutrial date string variables from tsExport data.}
\usage{
remove_st_dates_tsExport(tsExport, keep_vars = NULL)
}
\arguments{
\item{tsExport}{\code{tsExportdata} where original date variables should 
removed.}

\item{keep_vars}{character vector of variable names, that should be kept.}
}
\value{
\code{tsExportdata} without original secutrial date string variables.
}
\description{
Remove original secutrial date string variables, if an equivalent
date type converted variable exists. Remove the `.date` suffix from converted 
variable.
}
\seealso{
[remove_st_dates_tsExport_table()]
}
