% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{remove_st_dates_tsExport_table}
\alias{remove_st_dates_tsExport_table}
\title{Remove original secutrial date string variables from data table.}
\usage{
remove_st_dates_tsExport_table(.data, studydata_name, date_lookup)
}
\arguments{
\item{.data}{Data.Frame where the original secutrial date variables should be 
removed.}

\item{studydata_name}{String. Table name in lookup that relates to the table.}

\item{date_lookup}{List that holds date variables lookup information. It is 
produced by [create_date_lookup()]}
}
\value{
Data.Frame deleted secutrial date variables found in the vallab lookup 
list.
}
\description{
Remove original secutrial date string variables, if an equivalent
to date converted variable exists. Remove the `.date` suffix from converted 
variable.
}
\keyword{internal}
