% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{return_child_table_names_of_parent}
\alias{return_child_table_names_of_parent}
\title{Return parent table name (for one parent table)}
\usage{
return_child_table_names_of_parent(parentformtablename, trial_data)
}
\arguments{
\item{parentformtablename}{A vector with parent table name}

\item{trial_data}{A secuTrial data object}
}
\description{
Return parent table name (for one parent table)
}
