% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{return_children_table_names_of_parents}
\alias{return_children_table_names_of_parents}
\title{Return parent table names (for two or more parent tables)}
\usage{
return_children_table_names_of_parents(parentformtablenames, trial_data)
}
\arguments{
\item{parentformtablenames}{A vector with parent table names}

\item{trial_data}{A SecuTrial data object}
}
\description{
Return parent table names (for two or more parent tables)
}
