% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{return_parent_child_pairs}
\alias{return_parent_child_pairs}
\title{Return parent table names (for two or more parent tables)}
\usage{
return_parent_child_pairs(trial_data)
}
\arguments{
\item{trial_data}{A SecuTrial data object}
}
\description{
Return parent table names (for two or more parent tables)
}
