% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-id-names.R
\name{set_id_names}
\alias{set_id_names}
\title{Sets ID names and writes them into vectors}
\usage{
set_id_names(sT_or_ts_object)
}
\arguments{
\item{sT_or_ts_object}{data object}
}
\value{
sT_or_ts_object with additional value in export options
}
\description{
sets the name of the participant ID depending on sT or ts export.
}
