% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{spss_labelled_tsExport}
\alias{spss_labelled_tsExport}
\title{Add Value labels to tsExport data.}
\usage{
spss_labelled_tsExport(tsExport)
}
\arguments{
\item{tsExport}{\code{tsExportdata}.}
}
\value{
\code{tsExportdata} with variables in the vallab lookup list converted
to haven_labelled_spss class and their value labels assigned to them.
}
\description{
Convert numeric variables in the tsExport data that where found 
in the value labels lookup list into haven_labelled_spss class and
add value label attributes found in the lookup.
}
\seealso{
[valuelabels_tsExport()]
}
