% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{spss_labelled_tsExport_table}
\alias{spss_labelled_tsExport_table}
\title{Add value labels to tsExport data table.}
\usage{
spss_labelled_tsExport_table(.data, studydata_name, vallab_lookup)
}
\arguments{
\item{.data}{Data.Frame that should be spss_labelled.}

\item{studydata_name}{String. Table name in lookup that relates to the table
that get's spss_labelled.}

\item{vallab_lookup}{List that holds lookup information. It is produced by
[create_vallab_lookup()]}
}
\value{
Data.Frame with value labelled variables
}
\description{
Convert variables found in vallab lookup list into 
"haven_labelled_spss" class and set their `labels` attribute according to 
their entry in the lookup list.
}
\details{
This functions checks if all data frame variables found in the 
varlab lookup list are of haven_labelled class. If so, assume that they are 
already correctly labelled. If not apply [valuelabels_tsExport_table()] to 
label them. Afterwards add the `haven_labelled_spss` class attribute to those 
variables.
}
\seealso{
[valuelabels_tsExport_table()]
}
\keyword{internal}
