% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{spss_missings_tsExport}
\alias{spss_missings_tsExport}
\title{Set Missing Values in tsExport.}
\usage{
spss_missings_tsExport(
  tsExport,
  missing_labels = missing_labels_default,
  missing_values = NULL,
  overwrite_lookup = TRUE
)
}
\arguments{
\item{tsExport}{\code{tsExportdata}.}

\item{missing_labels}{character vector of value labels that are generally 
considered as missing values.}

\item{missing_values}{numerical vector of values that are generally considered
as missing values.}

\item{overwrite_lookup}{Boolean Should a new missing value lookup list created
when the tsExportdata object already contains one?}
}
\value{
\code{tsExportdata} with newly created missing value lookup list in
export_options and  `na_values` attributed variables.
}
\description{
Set missing value attribute according to the missing value
lookup list. If no lookup list exist in tha tsExport object or 
overwrite_lookup is TRUE create a new lookup list for for variables whose 
values match certain labels or numeric values that are generally 
considered as missing.
}
\details{
When a new missing value lookup list is created, it uses the value
label lookup list as foundation. The value lookup list get's filtered for 
matching labels or values that are provided as arguments to the function.  
As default, \code{missing_labels_default} is used to create the lookup list.
Afterwards variables found in the missing lookuplist, are converted to 
'haven_labelled_spss' class which uses the the `na_values` attribute to define
missing values. This attribute is set for each variable found in the 
missing value lookup list.
}
\seealso{
[create_missing_lookup()] [spss_missings_tsExport_table()]
}
