% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-SPSS-export.R
\name{spss_missings_tsExport_table}
\alias{spss_missings_tsExport_table}
\title{Set missing values to variables in a tsExport data table}
\usage{
spss_missings_tsExport_table(.data, studydata_name, missing_lookup)
}
\arguments{
\item{.data}{Data.Frame that should get missings attributed.}

\item{studydata_name}{String. Table name in lookup that relates to the table.}

\item{missing_lookup}{List that holds missing values lookup information. 
It is produced by [create_missing_lookup()].}
}
\value{
Data.Frame with `na_values` attributed haven_labelled_spss class 
variables.
}
\description{
Add missing value attribute to variables found in the missing 
values lookup list.
}
\details{
If a variable from the data frame is found in the missing value lookup list,
change the variable class to 'haven_labelled_spss' and set the `na_values` 
attribute to the values stored in the lookup.
}
\keyword{internal}
