% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{summarize_who_scale}
\alias{summarize_who_scale}
\title{Summarize WHO-Scale}
\usage{
summarize_who_scale(trial_data, pid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}
}
\value{
a vecctor, which contains the maximum who-scale per patient
}
\description{
Summarize WHO-Scale
}
