% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tsExport.R
\name{valuelabels_tsExport}
\alias{valuelabels_tsExport}
\title{Set value labels in NAPKON Transferstelle Export study data tables}
\usage{
valuelabels_tsExport(tsExport)
}
\arguments{
\item{tsExport}{\code{tsExportdata} that should get value labels}
}
\value{
\code{tsExportdata} with value labels
}
\description{
Set value labels for all variables found in the cl.csv file.

Create value_labels lookup list and add it to export_options.
}
\seealso{
valuelabels_tsExport_table
}
