% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tsExport.R
\name{valuelabels_tsExport_table}
\alias{valuelabels_tsExport_table}
\title{Set Value labels in a table}
\usage{
valuelabels_tsExport_table(.data, studydata_name, lookup)
}
\arguments{
\item{.data}{Data.Frame that should be labelled.}

\item{studydata_name}{String. Table name in lookup that relates to the table
that get's labelled.}

\item{lookup}{List that holds lookup information. It is produced by
\code{create_vallab_lookup}}
}
\value{
Data.Frame with (value) labelled variables
}
\description{
Set the attribute: labels in a tsExport data table.
}
\details{
If a variable is found in the lookup table, and it is already a labelled 
class vector, then just attach the labels attribute with the labels from the
lookup list.
If the variable is not a labelled class convert it with the labelled function
and attach value label function.
}
\keyword{internal}
