% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-tsExport.R
\name{write_tsExport}
\alias{write_tsExport}
\alias{write_tsExport.tsExportdata}
\alias{write_tsExport.data.frame}
\title{Write tsExport exports to other formats}
\usage{
write_tsExport(object, ...)

\method{write_tsExport}{tsExportdata}(object, format = "dta", metadata = FALSE, ...)

\method{write_tsExport}{data.frame}(df, filename, path = "", format = "dta", ...)
}
\arguments{
\item{object}{\code{tsExport} object}

\item{...}{further parameters}

\item{format}{format in which to save the export (one of "dta", "sas", "sav", "xpt")}

\item{metadata}{if TRUE then metadate files will also be written}

\item{df}{a data.frame}

\item{filename}{file name}

\item{path}{directory where the files should be saved}
}
\value{
a list of filenames
}
\description{
Convert the export prepared in R and export it to
             SPSS (sav), Stata (dta) or SAS (sas, xpt version 8)
             using the haven package.
}
\details{
Due to variable naming limitations in other packages, date variables are
         appended with _d (rather than _date), datetime/POSIX variables are appended
         with _dt (rather than _datetime) and factors with _cat (rather than _factor).
         Further variable names may be altered in the conversion process.
         For details please refer to the \code{haven} documentation.
         Since this has not been heavily tested or used there may be issues and you might prefer doing this manually
         with the haven package. One particular sticking point is the length of variable names - R is not restrictive in
         this respect, but other software can be. read_tsExport does not truncate names, prefering to leave this to the
         user, which can cause write_tsExport() to fail with an error.
}
\examples{
\dontrun{
# read ts data
EXPORT_DIR <- file.path("/path/to/transferstelle_export",
                        "export_directory")
# load all export data
tsExport <- read_tsExport(EXPORT_DIR, separator = ",", decimal = ".")
tdir <- tempdir()
write_tsExport(tsExport, format = "sav", path = tdir)
list.files(tdir)
}


}
\references{
Conversion to SPSS, STATA, SAS
https://github.com/SwissClinicalTrialOrganisation/secuTrialR/blob/master/R/write_secuTrial.R
https://cran.r-project.org/web/packages/secuTrialR/vignettes/secuTrialR-package-vignette.pdf
}
