% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_brs_n}
\alias{calculate_brs_n}
\title{Calculate Brief Resilience Scale (BRS) number of answered questions}
\usage{
calculate_brs_n(brs_1, brs_2, brs_3, brs_4, brs_5, brs_6)
}
\arguments{
\item{brs_1}{factor for item "Ich neige dazu, mich nach schwierigen Zeiten schnell zu erholen."}

\item{brs_2}{factor for item "Es fällt mir schwer, stressige Situationen durchzustehen."}

\item{brs_3}{factor for item "Ich brauche nicht viel Zeit, um mich von einem stressigen Ereignis zu erholen."}

\item{brs_4}{factor for item "Es fällt mir schwer, zur Normalität zurückzukehren, wenn etwas Schlimmes passiert ist."}

\item{brs_5}{factor for item "Normalerweise überstehe ich schwierige Zeiten ohne größere Probleme."}

\item{brs_6}{factor for item "Ich brauche tendenziell lange, um über Rückschläge in meinem Leben hinwegzukommen."}
}
\value{
A numeric vector with number of answered questions of BRS
}
\description{
Calculate Brief Resilience Scale (BRS) number of answered questions
}
