% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_brs_sum}
\alias{calculate_brs_sum}
\title{Calculate Brief Resilience Scale (BRS) sum score}
\usage{
calculate_brs_sum(brs_1, brs_2, brs_3, brs_4, brs_5, brs_6)
}
\arguments{
\item{brs_1}{factor for item "Ich neige dazu, mich nach schwierigen Zeiten schnell zu erholen."}

\item{brs_2}{factor for item "Es fällt mir schwer, stressige Situationen durchzustehen."}

\item{brs_3}{factor for item "Ich brauche nicht viel Zeit, um mich von einem stressigen Ereignis zu erholen."}

\item{brs_4}{factor for item "Es fällt mir schwer, zur Normalität zurückzukehren, wenn etwas Schlimmes passiert ist."}

\item{brs_5}{factor for item "Normalerweise überstehe ich schwierige Zeiten ohne größere Probleme."}

\item{brs_6}{factor for item "Ich brauche tendenziell lange, um über Rückschläge in meinem Leben hinwegzukommen."

Answers were coded /w levels "Stimme überhaupt nicht zu", "Stimme eher nicht zu", "Stimme eher zu", "Stimme vollkommen zu"}
}
\value{
A numeric vector with sum score of BRS
}
\description{
Calculate BRS sum score

Patients are asked "Bei den folgenden Fragen geht es um Ihre Widerstandsfähigkeit, Krisen zu bewältigen. 
Geben Sie bitte an, wie sehr Sie den folgenden Aussagen jeweils zustimmen."
}
