% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_facitf_sum}
\alias{calculate_facitf_sum}
\title{Calculate Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-F) sum score}
\usage{
calculate_facitf_sum(
  facitf_1,
  facitf_2,
  facitf_3,
  facitf_4,
  facitf_5,
  facitf_6,
  facitf_7,
  facitf_8,
  facitf_9,
  facitf_10,
  facitf_11,
  facitf_12,
  facitf_13
)
}
\arguments{
\item{facitf_1}{factor for item "Ich bin erschöpft."}

\item{facitf_2}{factor for item "Ich fühle mich insgesamt schwach."}

\item{facitf_3}{factor for item "Ich fühle mich lustlos (ausgelaugt)."}

\item{facitf_4}{factor for item "Ich bin müde."}

\item{facitf_5}{factor for item "Es fällt mir schwer, etwas anzufangen, weil ich müde bin."}

\item{facitf_6}{factor for item "Es fällt mir schwer, etwas zu Ende zu führen, weil ich müde bin."}

\item{facitf_7}{factor for item "Ich habe Energie."}

\item{facitf_8}{factor for item "Ich bin in der Lage, meinen gewohnten Aktivitäten nachzugehen (Beruf, Einkaufen, Schule, Freizeit, Sport usw.)."}

\item{facitf_9}{factor for item "Ich habe das Bedürfnis, tagsüber zu schlafen."}

\item{facitf_10}{factor for item "Ich bin zu müde, um zu essen."}

\item{facitf_11}{factor for item "Ich brauche Hilfe bei meinen gewohnten Aktivitäten (Beruf, Einkaufen, Schule, Freizeit, Sport usw.)."}

\item{facitf_12}{factor for item "Ich bin frustriert, weil ich zu müde bin, die Dinge zu tun, die ich machen möchte."}

\item{facitf_13}{factor for item "Ich muss meine sozialen Aktivitäten einschränken, weil ich müde bin."

Answers were coded /w levels "Überhaupt nicht", "Ein Wenig", "Mäßig", "Ziemlich" and "Sehr"}
}
\value{
A numeric vector with sum score of FACIT-F
}
\description{
Patients are asked "Bitte geben Sie jeweils an, wie sehr jede der folgenden Aussagen im Laufe der letzten 7 Tage auf Sie zugetroffen hat, 
indem Sie den entsprechenden Punkt auswählen."
}
