% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_full_years}
\alias{calculate_full_years}
\title{Primary coding of secutrial datasets of all NAPKON platforms
Primary coding functions developed for all NAPKON secutrial datasets.}
\usage{
calculate_full_years(from, to)
}
\arguments{
\item{from}{earlier date in yyyymmdd}

\item{to}{later date in yyyymmdd}
}
\value{
time difference between two dates in full years as numeric value
}
\description{
calculates time difference in full years
}
\details{
objective is to achieve a consistent and standardised preparation of variables: Categorization, definition of reference ranges, calculation of scores, etc. 
Calculate full years from two dates
}
