% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_gad7_sum}
\alias{calculate_gad7_sum}
\title{Calculate Generalized Anxiety Disorder 7 (GAD-7) sum score}
\usage{
calculate_gad7_sum(gad7_1, gad7_2, gad7_3, gad7_4, gad7_5, gad7_6, gad7_7)
}
\arguments{
\item{gad7_1}{factor for item "Nervosität, Ängstlichkeit oder Anspannung."}

\item{gad7_2}{factor for item "Nicht in der Lage sein, Sorgen zu stoppen oder zu kontrollieren."}

\item{gad7_3}{factor for item "Übermäßige Sorgen bezüglich verschiedener Angelegenheiten."}

\item{gad7_4}{factor for item "Schwierigkeiten zu entspannen."}

\item{gad7_5}{factor for item "Rastlosigkeit, so dass Stillsitzen schwer fällt."}

\item{gad7_6}{factor for item "Schnelle Verärgerung oder Gereiztheit."}

\item{gad7_7}{factor for item "Gefühl der Angst, so als würde etwas Schlimmes passieren."

Answers were coded /w levels 0 = "Überhaupt nicht", 1 = "An einzelnen Tagen", 2 = "An mehr als der Hälfte der Tage" and 3 = "Beinahme jeden Tag"}
}
\value{
A numeric vector with sum score of GAD-7
}
\description{
Calculate GAD-7 sum score

Patients are asked "Wie oft fühlten Sie sich im Verlauf der letzten 2 Wochen durch die folgenden Beschwerden beeinträchtigt?"
}
