% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_phq8_sum}
\alias{calculate_phq8_sum}
\title{Calculate Patient Health Questionnaire Depression Scale 8 (PHQ 8) sum score}
\usage{
calculate_phq8_sum(
  phq8_1,
  phq8_2,
  phq8_3,
  phq8_4,
  phq8_5,
  phq8_6,
  phq8_7,
  phq8_8
)
}
\arguments{
\item{phq8_1}{vector for item "Little interest or pleasure in doing things."}

\item{phq8_2}{vector for item "Feeling down, depressed, or hopeless."}

\item{phq8_3}{vector for item "Trouble falling or staying asleep, or sleeping too much."}

\item{phq8_4}{vector for item "Feeling tired or having little energy."}

\item{phq8_5}{vector for item "Poor appetite or overeating"}

\item{phq8_6}{vector for item "Feeling bad about yourself - or that you are a failure or have let yourself or your family down."}

\item{phq8_7}{vector for item "Trouble concentrating on things, such as reading the newspaper or watching television."}

\item{phq8_8}{vector for item "Moving or speaking so slowly that other people could have notived. Or the opposite - being so fidgety or restless that you have been moving around a lot more than usual."

#' Answers were coded /w levels 0 = "Not at all", 1 = "Several days", 2 = "More than half the days" and 3 = "Nearly every day"}
}
\value{
A numeric vector with sum score of phq8
}
\description{
Calculate sum score of PHQ-8

Patients are asked "Over the last 2 weeks, how often have you been bothered by any of the following problems?"
}
