% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{calculate_phq9_sum}
\alias{calculate_phq9_sum}
\title{Calculate Patient Health Questionnaire Depression Scale 9 (PHQ 9) sum score}
\usage{
calculate_phq9_sum(
  phq9_1,
  phq9_2,
  phq9_3,
  phq9_4,
  phq9_5,
  phq9_6,
  phq9_7,
  phq9_8,
  phq9_9
)
}
\arguments{
\item{phq9_1}{vector for item "Little interest or pleasure in doing things."}

\item{phq9_2}{vector for item "Feeling down, depressed, or hopeless."}

\item{phq9_3}{vector for item "Trouble falling or staying asleep, or sleeping too much."}

\item{phq9_4}{vector for item "Feeling tired or having little energy."}

\item{phq9_5}{vector for item "Poor appetite or overeating"}

\item{phq9_6}{vector for item "Feeling bad about yourself - or that you are a failure or have let yourself or your family down."}

\item{phq9_7}{vector for item "Trouble concentrating on things, such as reading the newspaper or watching television."}

\item{phq9_8}{vector for item "Moving or speaking so slowly that other people could have notived. Or the opposite - being so fidgety or restless that you have been moving around a lot more than usual."}

\item{phq9_9}{vector for item "Thoughts that you would be better off dead, or thoughts of hurting yourself in some way?"

Answers were coded /w levels 0 = "Not at all", 1 = "Several days", 2 = "More than half the days" and 3 = "Nearly every day"}
}
\value{
A numeric vector with sum score of phq9
}
\description{
Calculate sum score of PHQ-8

Patients are asked "Over the last 2 weeks, how often have you been bothered by any of the following problems?"
}
