% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding.R
\name{categorize_facitf_ecu}
\alias{categorize_facitf_ecu}
\title{Categorize Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-F)}
\usage{
categorize_facitf_ecu(ecu_facitf_sum)
}
\arguments{
\item{ecu_facitf_sum}{numeric vector with sum score of FACIT-F}
}
\value{
A factorized vector w/ levels "Relevant fatigue" and "No fatigue"
}
\description{
Categorize FACIT-F sum score in no fatigue and relevant fatigue
}
