% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_apache2}
\alias{primary_coding_hap_apache2}
\title{Primary coding for Acute Physiology And Chronic Health Evaluation (APACHE) Score}
\usage{
primary_coding_hap_apache2(trial_data, table_names)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{table_names}{A vector containing table names of data export}
}
\description{
adds the following column to klinscores1:
ecu_apache2_cat
}
