% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_barthel_disc}
\alias{primary_coding_hap_barthel_disc}
\title{Primary coding Barthel index at discharge}
\usage{
primary_coding_hap_barthel_disc(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to end:
ecu_barthel_cat_disc
}
