% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_barthel_pre}
\alias{primary_coding_hap_barthel_pre}
\title{Primary coding Barthel index prior to infection (baseline)}
\usage{
primary_coding_hap_barthel_pre(trial_data, table_names)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{table_names}{A vector containing table names of data export}
}
\description{
adds the following column to risiko1:
ecu_barthel_cat_pre
}
