% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_bmi}
\alias{primary_coding_hap_bmi}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_hap_bmi(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following columns to _demo: ecu_bmi, ecu_bmi_cat
}
