% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_clinical_params}
\alias{primary_coding_hap_clinical_params}
\title{Primary coding clinical parameters}
\usage{
primary_coding_hap_clinical_params(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to vitalparam: 
ecu_vitals_bp, ecu_vitals_bpm, ecu_vitals_so2, ecu_vitals_resp_rate, ecu_vitals_temp,
ecu_vitals_gcs, ecu_vitals_ph
}
