% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_dds}
\alias{primary_coding_hap_dds}
\title{Primary coding for Delirium Detection Score (DDS)}
\usage{
primary_coding_hap_dds(trial_data, table_names)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{table_names}{A vector containing table names of data export}
}
\description{
adds the following column to haemodyn:
ecu_dds_cat
}
