% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_eq5d5l}
\alias{primary_coding_hap_eq5d5l}
\title{Primary coding EQ5D-5L-Index}
\usage{
primary_coding_hap_eq5d5l(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to _eq5d:
ecu_eq5d_index
}
