% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_icdsc}
\alias{primary_coding_hap_icdsc}
\title{Primary coding for Intensive Care Delirium Screening Checklist (ICDSC)}
\usage{
primary_coding_hap_icdsc(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to _haemodyn:
ecu_icdsc_cat
}
