% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-hap.R
\name{primary_coding_hap_news_first}
\alias{primary_coding_hap_news_first}
\title{Primary coding for National Early Warning Score (NEWS)}
\usage{
primary_coding_hap_news_first(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to _klinscores:
ecu_news_cat
}
