% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_abd_overw}
\alias{primary_coding_pop_abd_overw}
\title{Primary coding Body Mass Index (BMI)}
\usage{
primary_coding_pop_abd_overw(trial_data, pid)
}
\arguments{
\item{trial_data}{A secuTrial data object}

\item{pid}{column name of patient ID in trial_data}
}
\description{
adds the following columns to anthropo: 
ecu_waist
}
