% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_facitf}
\alias{primary_coding_pop_facitf}
\title{Primary coding Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-F)}
\usage{
primary_coding_pop_facitf(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to surveyfrageboge: 
ecu_facitf
}
