% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_facitf}
\alias{primary_coding_pop_facitf}
\title{Primary coding Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-Fatigue Scale)}
\usage{
primary_coding_pop_facitf(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to surveyfrageboge: 
ecu_facitf_sum, ecu_facitf_n, ecu_facitf_scale, ecu_facitf_cat
ecu_facitf1, ecu_facitf2, ecu_facitf3, ecu_facitf4, ecu_facitf5, ecu_facitf6, ecu_facitf7, ecu_facitf8, ecu_facitf8
ecu_facitf10, ecu_facitf11, ecu_facitf12, ecu_facitf13
}
