% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-pop.R
\name{primary_coding_pop_gad7}
\alias{primary_coding_pop_gad7}
\title{Primary coding Generalized Anxiety Disorder 7 (GAD-7)}
\usage{
primary_coding_pop_gad7(trial_data)
}
\arguments{
\item{trial_data}{SecuTrial Data}
}
\description{
adds the following variable to surveyfragebogen: 
ecu_gad7
}
