% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-rapid_revive.R
\name{primary_coding_rapid_revive_sf36}
\alias{primary_coding_rapid_revive_sf36}
\title{Short Form Survey Instrument (SF-36) questionnaire}
\usage{
primary_coding_rapid_revive_sf36(trial_data)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following column to sf36: 
ecu_sf_013, ecu_sf_014, ecu_sf_015, ecu_sf_016. ecu_sf_017, v18, ecu_sf_019 (recoded items)
ecu_sf_001p, ecu_sf_002p, ecu_sf_003p, ecu_sf_004p, ecu_sf_005p, ecu_sf_006p, ecu_sf_007p, ecu_sf_008p, ecu_sf_009p,
ecu_sf_010p, ecu_sf_011p, ecu_sf_012p, ecu_sf_013p, ecu_sf_014p, ecu_sf_015p, ecu_sf_016p, ecu_sf_017p, ecu_sf_018p,
ecu_sf_019p, ecu_sf_020p, ecu_sf_021p, ecu_sf_022p, ecu_sf_023p, ecu_sf_024p, ecu_sf_025p, ecu_sf_026p, ecu_sf_027p,
ecu_sf_028p, ecu_sf_029p, ecu_sf_030p, ecu_sf_031p, ecu_sf_032p, ecu_sf_033p, ecu_sf_034p, ecu_sf_035p, ecu_sf_036p (points)
ecu_sf_phys, ecu_sf_rl_phys, ecu_sf_rl_emot, ecu_sf_vitality, ecu_sf_wb, ecu_sf_social, ecu_sf_pain, ecu_sf_general,
ecu_sf_hc (mean values)
}
