% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-coding-suep.R
\name{primary_coding_suep_brs}
\alias{primary_coding_suep_brs}
\title{Primary coding Brief Resilience Scale (BRS)}
\usage{
primary_coding_suep_brs(trial_data, pid, visitid)
}
\arguments{
\item{trial_data}{A secuTrial data object}
}
\description{
adds the following columns to promext: 
ecu_brs_sum, ecu_brs_n, ecu_brs_total, ecu_brs_cat
}
